Embed remote sockjs-client js and minify vaadinPushSockJS
