--- vertx-vaadin/src/main/resources/VAADIN/vaadinPushSockJS.js
@@ -1,6257 +1,0 @@
-window.vaadinPush = window.vaadinPush || {};
-(function(define) {
-
-  /* sockjs-client v1.1.4 | http://sockjs.org | MIT license */
-  (function(f) {
-    if (typeof exports === "object" && typeof module !== "undefined") {
-      module.exports = f()
-    } else if (typeof define === "function" && define.amd) {
-      define([], f)
-    } else {
-      var g;
-      if (typeof window !== "undefined") {
-        g = window
-      } else if (typeof global !== "undefined") {
-        g = global
-      } else if (typeof self !== "undefined") {
-        g = self
-      } else {
-        g = this
-      }
-      g.SockJS = f()
-    }
-  })(function() {
-    var define, module, exports;
-    return (function e(t, n, r) {
-      function s(o, u) {
-        if (!n[o]) {
-          if (!t[o]) {
-            var a = typeof require == "function" && require;
-            if (!u && a) return a(o, !0);
-            if (i) return i(o, !0);
-            var f = new Error("Cannot find module '" + o + "'");
-            throw f.code = "MODULE_NOT_FOUND", f
-          }
-          var l = n[o] = {
-            exports: {}
-          };
-          t[o][0].call(l.exports, function(e) {
-            var n = t[o][1][e];
-            return s(n ? n : e)
-          }, l, l.exports, e, t, n, r)
-        }
-        return n[o].exports
-      }
-      var i = typeof require == "function" && require;
-      for (var o = 0; o < r.length; o++) s(r[o]);
-      return s
-    })({
-      1: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var transportList = require('./transport-list');
-
-          module.exports = require('./main')(transportList);
-
-          // TODO can't get rid of this until all servers do
-          if ('_sockjs_onload' in global) {
-            setTimeout(global._sockjs_onload, 1);
-          }
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./main": 14,
-        "./transport-list": 16
-      }],
-      2: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          Event = require('./event');
-
-        function CloseEvent() {
-          Event.call(this);
-          this.initEvent('close', false, false);
-          this.wasClean = false;
-          this.code = 0;
-          this.reason = '';
-        }
-
-        inherits(CloseEvent, Event);
-
-        module.exports = CloseEvent;
-
-      }, {
-        "./event": 4,
-        "inherits": 57
-      }],
-      3: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          EventTarget = require('./eventtarget');
-
-        function EventEmitter() {
-          EventTarget.call(this);
-        }
-
-        inherits(EventEmitter, EventTarget);
-
-        EventEmitter.prototype.removeAllListeners = function(type) {
-          if (type) {
-            delete this._listeners[type];
-          } else {
-            this._listeners = {};
-          }
-        };
-
-        EventEmitter.prototype.once = function(type, listener) {
-          var self = this,
-            fired = false;
-
-          function g() {
-            self.removeListener(type, g);
-
-            if (!fired) {
-              fired = true;
-              listener.apply(this, arguments);
-            }
-          }
-
-          this.on(type, g);
-        };
-
-        EventEmitter.prototype.emit = function() {
-          var type = arguments[0];
-          var listeners = this._listeners[type];
-          if (!listeners) {
-            return;
-          }
-          // equivalent of Array.prototype.slice.call(arguments, 1);
-          var l = arguments.length;
-          var args = new Array(l - 1);
-          for (var ai = 1; ai < l; ai++) {
-            args[ai - 1] = arguments[ai];
-          }
-          for (var i = 0; i < listeners.length; i++) {
-            listeners[i].apply(this, args);
-          }
-        };
-
-        EventEmitter.prototype.on = EventEmitter.prototype.addListener = EventTarget.prototype.addEventListener;
-        EventEmitter.prototype.removeListener = EventTarget.prototype.removeEventListener;
-
-        module.exports.EventEmitter = EventEmitter;
-
-      }, {
-        "./eventtarget": 5,
-        "inherits": 57
-      }],
-      4: [function(require, module, exports) {
-        'use strict';
-
-        function Event(eventType) {
-          this.type = eventType;
-        }
-
-        Event.prototype.initEvent = function(eventType, canBubble, cancelable) {
-          this.type = eventType;
-          this.bubbles = canBubble;
-          this.cancelable = cancelable;
-          this.timeStamp = +new Date();
-          return this;
-        };
-
-        Event.prototype.stopPropagation = function() {};
-        Event.prototype.preventDefault = function() {};
-
-        Event.CAPTURING_PHASE = 1;
-        Event.AT_TARGET = 2;
-        Event.BUBBLING_PHASE = 3;
-
-        module.exports = Event;
-
-      }, {}],
-      5: [function(require, module, exports) {
-        'use strict';
-
-        /* Simplified implementation of DOM2 EventTarget.
-         *   http://www.w3.org/TR/DOM-Level-2-Events/events.html#Events-EventTarget
-         */
-
-        function EventTarget() {
-          this._listeners = {};
-        }
-
-        EventTarget.prototype.addEventListener = function(eventType, listener) {
-          if (!(eventType in this._listeners)) {
-            this._listeners[eventType] = [];
-          }
-          var arr = this._listeners[eventType];
-          // #4
-          if (arr.indexOf(listener) === -1) {
-            // Make a copy so as not to interfere with a current dispatchEvent.
-            arr = arr.concat([listener]);
-          }
-          this._listeners[eventType] = arr;
-        };
-
-        EventTarget.prototype.removeEventListener = function(eventType, listener) {
-          var arr = this._listeners[eventType];
-          if (!arr) {
-            return;
-          }
-          var idx = arr.indexOf(listener);
-          if (idx !== -1) {
-            if (arr.length > 1) {
-              // Make a copy so as not to interfere with a current dispatchEvent.
-              this._listeners[eventType] = arr.slice(0, idx).concat(arr.slice(idx + 1));
-            } else {
-              delete this._listeners[eventType];
-            }
-            return;
-          }
-        };
-
-        EventTarget.prototype.dispatchEvent = function() {
-          var event = arguments[0];
-          var t = event.type;
-          // equivalent of Array.prototype.slice.call(arguments, 0);
-          var args = arguments.length === 1 ? [event] : Array.apply(null, arguments);
-          // TODO: This doesn't match the real behavior; per spec, onfoo get
-          // their place in line from the /first/ time they're set from
-          // non-null. Although WebKit bumps it to the end every time it's
-          // set.
-          if (this['on' + t]) {
-            this['on' + t].apply(this, args);
-          }
-          if (t in this._listeners) {
-            // Grab a reference to the listeners list. removeEventListener may alter the list.
-            var listeners = this._listeners[t];
-            for (var i = 0; i < listeners.length; i++) {
-              listeners[i].apply(this, args);
-            }
-          }
-        };
-
-        module.exports = EventTarget;
-
-      }, {}],
-      6: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          Event = require('./event');
-
-        function TransportMessageEvent(data) {
-          Event.call(this);
-          this.initEvent('message', false, false);
-          this.data = data;
-        }
-
-        inherits(TransportMessageEvent, Event);
-
-        module.exports = TransportMessageEvent;
-
-      }, {
-        "./event": 4,
-        "inherits": 57
-      }],
-      7: [function(require, module, exports) {
-        'use strict';
-
-        var JSON3 = require('json3'),
-          iframeUtils = require('./utils/iframe');
-
-        function FacadeJS(transport) {
-          this._transport = transport;
-          transport.on('message', this._transportMessage.bind(this));
-          transport.on('close', this._transportClose.bind(this));
-        }
-
-        FacadeJS.prototype._transportClose = function(code, reason) {
-          iframeUtils.postMessage('c', JSON3.stringify([code, reason]));
-        };
-        FacadeJS.prototype._transportMessage = function(frame) {
-          iframeUtils.postMessage('t', frame);
-        };
-        FacadeJS.prototype._send = function(data) {
-          this._transport.send(data);
-        };
-        FacadeJS.prototype._close = function() {
-          this._transport.close();
-          this._transport.removeAllListeners();
-        };
-
-        module.exports = FacadeJS;
-
-      }, {
-        "./utils/iframe": 47,
-        "json3": 58
-      }],
-      8: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var urlUtils = require('./utils/url'),
-            eventUtils = require('./utils/event'),
-            JSON3 = require('json3'),
-            FacadeJS = require('./facade'),
-            InfoIframeReceiver = require('./info-iframe-receiver'),
-            iframeUtils = require('./utils/iframe'),
-            loc = require('./location');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:iframe-bootstrap');
-          }
-
-          module.exports = function(SockJS, availableTransports) {
-            var transportMap = {};
-            availableTransports.forEach(function(at) {
-              if (at.facadeTransport) {
-                transportMap[at.facadeTransport.transportName] = at.facadeTransport;
-              }
-            });
-
-            // hard-coded for the info iframe
-            // TODO see if we can make this more dynamic
-            transportMap[InfoIframeReceiver.transportName] = InfoIframeReceiver;
-            var parentOrigin;
-
-            /* eslint-disable camelcase */
-            SockJS.bootstrap_iframe = function() {
-              /* eslint-enable camelcase */
-              var facade;
-              iframeUtils.currentWindowId = loc.hash.slice(1);
-              var onMessage = function(e) {
-                if (e.source !== parent) {
-                  return;
-                }
-                if (typeof parentOrigin === 'undefined') {
-                  parentOrigin = e.origin;
-                }
-                if (e.origin !== parentOrigin) {
-                  return;
-                }
-
-                var iframeMessage;
-                try {
-                  iframeMessage = JSON3.parse(e.data);
-                } catch (ignored) {
-                  debug('bad json', e.data);
-                  return;
-                }
-
-                if (iframeMessage.windowId !== iframeUtils.currentWindowId) {
-                  return;
-                }
-                switch (iframeMessage.type) {
-                  case 's':
-                    var p;
-                    try {
-                      p = JSON3.parse(iframeMessage.data);
-                    } catch (ignored) {
-                      debug('bad json', iframeMessage.data);
-                      break;
-                    }
-                    var version = p[0];
-                    var transport = p[1];
-                    var transUrl = p[2];
-                    var baseUrl = p[3];
-                    debug(version, transport, transUrl, baseUrl);
-                    // change this to semver logic
-                    if (version !== SockJS.version) {
-                      throw new Error('Incompatible SockJS! Main site uses:' +
-                        ' "' + version + '", the iframe:' +
-                        ' "' + SockJS.version + '".');
-                    }
-
-                    if (!urlUtils.isOriginEqual(transUrl, loc.href) ||
-                      !urlUtils.isOriginEqual(baseUrl, loc.href)) {
-                      throw new Error('Can\'t connect to different domain from within an ' +
-                        'iframe. (' + loc.href + ', ' + transUrl + ', ' + baseUrl + ')');
-                    }
-                    facade = new FacadeJS(new transportMap[transport](transUrl, baseUrl));
-                    break;
-                  case 'm':
-                    facade._send(iframeMessage.data);
-                    break;
-                  case 'c':
-                    if (facade) {
-                      facade._close();
-                    }
-                    facade = null;
-                    break;
-                }
-              };
-
-              eventUtils.attachEvent('message', onMessage);
-
-              // Start
-              iframeUtils.postMessage('s');
-            };
-          };
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "./facade": 7,
-        "./info-iframe-receiver": 10,
-        "./location": 13,
-        "./utils/event": 46,
-        "./utils/iframe": 47,
-        "./utils/url": 52,
-        "debug": 55,
-        "json3": 58
-      }],
-      9: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var EventEmitter = require('events').EventEmitter,
-            inherits = require('inherits'),
-            JSON3 = require('json3'),
-            objectUtils = require('./utils/object');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:info-ajax');
-          }
-
-          function InfoAjax(url, AjaxObject) {
-            EventEmitter.call(this);
-
-            var self = this;
-            var t0 = +new Date();
-            this.xo = new AjaxObject('GET', url);
-
-            this.xo.once('finish', function(status, text) {
-              var info, rtt;
-              if (status === 200) {
-                rtt = (+new Date()) - t0;
-                if (text) {
-                  try {
-                    info = JSON3.parse(text);
-                  } catch (e) {
-                    debug('bad json', text);
-                  }
-                }
-
-                if (!objectUtils.isObject(info)) {
-                  info = {};
-                }
-              }
-              self.emit('finish', info, rtt);
-              self.removeAllListeners();
-            });
-          }
-
-          inherits(InfoAjax, EventEmitter);
-
-          InfoAjax.prototype.close = function() {
-            this.removeAllListeners();
-            this.xo.close();
-          };
-
-          module.exports = InfoAjax;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "./utils/object": 49,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57,
-        "json3": 58
-      }],
-      10: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          EventEmitter = require('events').EventEmitter,
-          JSON3 = require('json3'),
-          XHRLocalObject = require('./transport/sender/xhr-local'),
-          InfoAjax = require('./info-ajax');
-
-        function InfoReceiverIframe(transUrl) {
-          var self = this;
-          EventEmitter.call(this);
-
-          this.ir = new InfoAjax(transUrl, XHRLocalObject);
-          this.ir.once('finish', function(info, rtt) {
-            self.ir = null;
-            self.emit('message', JSON3.stringify([info, rtt]));
-          });
-        }
-
-        inherits(InfoReceiverIframe, EventEmitter);
-
-        InfoReceiverIframe.transportName = 'iframe-info-receiver';
-
-        InfoReceiverIframe.prototype.close = function() {
-          if (this.ir) {
-            this.ir.close();
-            this.ir = null;
-          }
-          this.removeAllListeners();
-        };
-
-        module.exports = InfoReceiverIframe;
-
-      }, {
-        "./info-ajax": 9,
-        "./transport/sender/xhr-local": 37,
-        "events": 3,
-        "inherits": 57,
-        "json3": 58
-      }],
-      11: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var EventEmitter = require('events').EventEmitter,
-            inherits = require('inherits'),
-            JSON3 = require('json3'),
-            utils = require('./utils/event'),
-            IframeTransport = require('./transport/iframe'),
-            InfoReceiverIframe = require('./info-iframe-receiver');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:info-iframe');
-          }
-
-          function InfoIframe(baseUrl, url) {
-            var self = this;
-            EventEmitter.call(this);
-
-            var go = function() {
-              var ifr = self.ifr = new IframeTransport(InfoReceiverIframe.transportName, url, baseUrl);
-
-              ifr.once('message', function(msg) {
-                if (msg) {
-                  var d;
-                  try {
-                    d = JSON3.parse(msg);
-                  } catch (e) {
-                    debug('bad json', msg);
-                    self.emit('finish');
-                    self.close();
-                    return;
-                  }
-
-                  var info = d[0],
-                    rtt = d[1];
-                  self.emit('finish', info, rtt);
-                }
-                self.close();
-              });
-
-              ifr.once('close', function() {
-                self.emit('finish');
-                self.close();
-              });
-            };
-
-            // TODO this seems the same as the 'needBody' from transports
-            if (!global.document.body) {
-              utils.attachEvent('load', go);
-            } else {
-              go();
-            }
-          }
-
-          inherits(InfoIframe, EventEmitter);
-
-          InfoIframe.enabled = function() {
-            return IframeTransport.enabled();
-          };
-
-          InfoIframe.prototype.close = function() {
-            if (this.ifr) {
-              this.ifr.close();
-            }
-            this.removeAllListeners();
-            this.ifr = null;
-          };
-
-          module.exports = InfoIframe;
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./info-iframe-receiver": 10,
-        "./transport/iframe": 22,
-        "./utils/event": 46,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57,
-        "json3": 58
-      }],
-      12: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var EventEmitter = require('events').EventEmitter,
-            inherits = require('inherits'),
-            urlUtils = require('./utils/url'),
-            XDR = require('./transport/sender/xdr'),
-            XHRCors = require('./transport/sender/xhr-cors'),
-            XHRLocal = require('./transport/sender/xhr-local'),
-            XHRFake = require('./transport/sender/xhr-fake'),
-            InfoIframe = require('./info-iframe'),
-            InfoAjax = require('./info-ajax');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:info-receiver');
-          }
-
-          function InfoReceiver(baseUrl, urlInfo) {
-            debug(baseUrl);
-            var self = this;
-            EventEmitter.call(this);
-
-            setTimeout(function() {
-              self.doXhr(baseUrl, urlInfo);
-            }, 0);
-          }
-
-          inherits(InfoReceiver, EventEmitter);
-
-          // TODO this is currently ignoring the list of available transports and the whitelist
-
-          InfoReceiver._getReceiver = function(baseUrl, url, urlInfo) {
-            // determine method of CORS support (if needed)
-            if (urlInfo.sameOrigin) {
-              return new InfoAjax(url, XHRLocal);
-            }
-            if (XHRCors.enabled) {
-              return new InfoAjax(url, XHRCors);
-            }
-            if (XDR.enabled && urlInfo.sameScheme) {
-              return new InfoAjax(url, XDR);
-            }
-            if (InfoIframe.enabled()) {
-              return new InfoIframe(baseUrl, url);
-            }
-            return new InfoAjax(url, XHRFake);
-          };
-
-          InfoReceiver.prototype.doXhr = function(baseUrl, urlInfo) {
-            var self = this,
-              url = urlUtils.addPath(baseUrl, '/info');
-            debug('doXhr', url);
-
-            this.xo = InfoReceiver._getReceiver(baseUrl, url, urlInfo);
-
-            this.timeoutRef = setTimeout(function() {
-              debug('timeout');
-              self._cleanup(false);
-              self.emit('finish');
-            }, InfoReceiver.timeout);
-
-            this.xo.once('finish', function(info, rtt) {
-              debug('finish', info, rtt);
-              self._cleanup(true);
-              self.emit('finish', info, rtt);
-            });
-          };
-
-          InfoReceiver.prototype._cleanup = function(wasClean) {
-            debug('_cleanup');
-            clearTimeout(this.timeoutRef);
-            this.timeoutRef = null;
-            if (!wasClean && this.xo) {
-              this.xo.close();
-            }
-            this.xo = null;
-          };
-
-          InfoReceiver.prototype.close = function() {
-            debug('close');
-            this.removeAllListeners();
-            this._cleanup(false);
-          };
-
-          InfoReceiver.timeout = 8000;
-
-          module.exports = InfoReceiver;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "./info-ajax": 9,
-        "./info-iframe": 11,
-        "./transport/sender/xdr": 34,
-        "./transport/sender/xhr-cors": 35,
-        "./transport/sender/xhr-fake": 36,
-        "./transport/sender/xhr-local": 37,
-        "./utils/url": 52,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      13: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          module.exports = global.location || {
-            origin: 'http://localhost:80',
-            protocol: 'http',
-            host: 'localhost',
-            port: 80,
-            href: 'http://localhost/',
-            hash: ''
-          };
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      14: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          require('./shims');
-
-          var URL = require('url-parse'),
-            inherits = require('inherits'),
-            JSON3 = require('json3'),
-            random = require('./utils/random'),
-            escape = require('./utils/escape'),
-            urlUtils = require('./utils/url'),
-            eventUtils = require('./utils/event'),
-            transport = require('./utils/transport'),
-            objectUtils = require('./utils/object'),
-            browser = require('./utils/browser'),
-            log = require('./utils/log'),
-            Event = require('./event/event'),
-            EventTarget = require('./event/eventtarget'),
-            loc = require('./location'),
-            CloseEvent = require('./event/close'),
-            TransportMessageEvent = require('./event/trans-message'),
-            InfoReceiver = require('./info-receiver');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:main');
-          }
-
-          var transports;
-
-          // follow constructor steps defined at http://dev.w3.org/html5/websockets/#the-websocket-interface
-          function SockJS(url, protocols, options) {
-            if (!(this instanceof SockJS)) {
-              return new SockJS(url, protocols, options);
-            }
-            if (arguments.length < 1) {
-              throw new TypeError("Failed to construct 'SockJS: 1 argument required, but only 0 present");
-            }
-            EventTarget.call(this);
-
-            this.readyState = SockJS.CONNECTING;
-            this.extensions = '';
-            this.protocol = '';
-
-            // non-standard extension
-            options = options || {};
-            if (options.protocols_whitelist) {
-              log.warn("'protocols_whitelist' is DEPRECATED. Use 'transports' instead.");
-            }
-            this._transportsWhitelist = options.transports;
-            this._transportOptions = options.transportOptions || {};
-
-            var sessionId = options.sessionId || 8;
-            if (typeof sessionId === 'function') {
-              this._generateSessionId = sessionId;
-            } else if (typeof sessionId === 'number') {
-              this._generateSessionId = function() {
-                return random.string(sessionId);
-              };
-            } else {
-              throw new TypeError('If sessionId is used in the options, it needs to be a number or a function.');
-            }
-
-            this._server = options.server || random.numberString(1000);
-
-            // Step 1 of WS spec - parse and validate the url. Issue #8
-            var parsedUrl = new URL(url);
-            if (!parsedUrl.host || !parsedUrl.protocol) {
-              throw new SyntaxError("The URL '" + url + "' is invalid");
-            } else if (parsedUrl.hash) {
-              throw new SyntaxError('The URL must not contain a fragment');
-            } else if (parsedUrl.protocol !== 'http:' && parsedUrl.protocol !== 'https:') {
-              throw new SyntaxError("The URL's scheme must be either 'http:' or 'https:'. '" + parsedUrl.protocol + "' is not allowed.");
-            }
-
-            var secure = parsedUrl.protocol === 'https:';
-            // Step 2 - don't allow secure origin with an insecure protocol
-            if (loc.protocol === 'https' && !secure) {
-              throw new Error('SecurityError: An insecure SockJS connection may not be initiated from a page loaded over HTTPS');
-            }
-
-            // Step 3 - check port access - no need here
-            // Step 4 - parse protocols argument
-            if (!protocols) {
-              protocols = [];
-            } else if (!Array.isArray(protocols)) {
-              protocols = [protocols];
-            }
-
-            // Step 5 - check protocols argument
-            var sortedProtocols = protocols.sort();
-            sortedProtocols.forEach(function(proto, i) {
-              if (!proto) {
-                throw new SyntaxError("The protocols entry '" + proto + "' is invalid.");
-              }
-              if (i < (sortedProtocols.length - 1) && proto === sortedProtocols[i + 1]) {
-                throw new SyntaxError("The protocols entry '" + proto + "' is duplicated.");
-              }
-            });
-
-            // Step 6 - convert origin
-            var o = urlUtils.getOrigin(loc.href);
-            this._origin = o ? o.toLowerCase() : null;
-
-            // remove the trailing slash
-            parsedUrl.set('pathname', parsedUrl.pathname.replace(/\/+$/, ''));
-
-            // store the sanitized url
-            this.url = parsedUrl.href;
-            debug('using url', this.url);
-
-            // Step 7 - start connection in background
-            // obtain server info
-            // http://sockjs.github.io/sockjs-protocol/sockjs-protocol-0.3.3.html#section-26
-            this._urlInfo = {
-              nullOrigin: !browser.hasDomain(),
-              sameOrigin: urlUtils.isOriginEqual(this.url, loc.href),
-              sameScheme: urlUtils.isSchemeEqual(this.url, loc.href)
-            };
-
-            this._ir = new InfoReceiver(this.url, this._urlInfo);
-            this._ir.once('finish', this._receiveInfo.bind(this));
-          }
-
-          inherits(SockJS, EventTarget);
-
-          function userSetCode(code) {
-            return code === 1000 || (code >= 3000 && code <= 4999);
-          }
-
-          SockJS.prototype.close = function(code, reason) {
-            // Step 1
-            if (code && !userSetCode(code)) {
-              throw new Error('InvalidAccessError: Invalid code');
-            }
-            // Step 2.4 states the max is 123 bytes, but we are just checking length
-            if (reason && reason.length > 123) {
-              throw new SyntaxError('reason argument has an invalid length');
-            }
-
-            // Step 3.1
-            if (this.readyState === SockJS.CLOSING || this.readyState === SockJS.CLOSED) {
-              return;
-            }
-
-            // TODO look at docs to determine how to set this
-            var wasClean = true;
-            this._close(code || 1000, reason || 'Normal closure', wasClean);
-          };
-
-          SockJS.prototype.send = function(data) {
-            // #13 - convert anything non-string to string
-            // TODO this currently turns objects into [object Object]
-            if (typeof data !== 'string') {
-              data = '' + data;
-            }
-            if (this.readyState === SockJS.CONNECTING) {
-              throw new Error('InvalidStateError: The connection has not been established yet');
-            }
-            if (this.readyState !== SockJS.OPEN) {
-              return;
-            }
-            this._transport.send(escape.quote(data));
-          };
-
-          SockJS.version = require('./version');
-
-          SockJS.CONNECTING = 0;
-          SockJS.OPEN = 1;
-          SockJS.CLOSING = 2;
-          SockJS.CLOSED = 3;
-
-          SockJS.prototype._receiveInfo = function(info, rtt) {
-            debug('_receiveInfo', rtt);
-            this._ir = null;
-            if (!info) {
-              this._close(1002, 'Cannot connect to server');
-              return;
-            }
-
-            // establish a round-trip timeout (RTO) based on the
-            // round-trip time (RTT)
-            this._rto = this.countRTO(rtt);
-            // allow server to override url used for the actual transport
-            this._transUrl = info.base_url ? info.base_url : this.url;
-            info = objectUtils.extend(info, this._urlInfo);
-            debug('info', info);
-            // determine list of desired and supported transports
-            var enabledTransports = transports.filterToEnabled(this._transportsWhitelist, info);
-            this._transports = enabledTransports.main;
-            debug(this._transports.length + ' enabled transports');
-
-            this._connect();
-          };
-
-          SockJS.prototype._connect = function() {
-            for (var Transport = this._transports.shift(); Transport; Transport = this._transports.shift()) {
-              debug('attempt', Transport.transportName);
-              if (Transport.needBody) {
-                if (!global.document.body ||
-                  (typeof global.document.readyState !== 'undefined' &&
-                    global.document.readyState !== 'complete' &&
-                    global.document.readyState !== 'interactive')) {
-                  debug('waiting for body');
-                  this._transports.unshift(Transport);
-                  eventUtils.attachEvent('load', this._connect.bind(this));
-                  return;
-                }
-              }
-
-              // calculate timeout based on RTO and round trips. Default to 5s
-              var timeoutMs = (this._rto * Transport.roundTrips) || 5000;
-              this._transportTimeoutId = setTimeout(this._transportTimeout.bind(this), timeoutMs);
-              debug('using timeout', timeoutMs);
-
-              var transportUrl = urlUtils.addPath(this._transUrl, '/' + this._server + '/' + this._generateSessionId());
-              var options = this._transportOptions[Transport.transportName];
-              debug('transport url', transportUrl);
-              var transportObj = new Transport(transportUrl, this._transUrl, options);
-              transportObj.on('message', this._transportMessage.bind(this));
-              transportObj.once('close', this._transportClose.bind(this));
-              transportObj.transportName = Transport.transportName;
-              this._transport = transportObj;
-
-              return;
-            }
-            this._close(2000, 'All transports failed', false);
-          };
-
-          SockJS.prototype._transportTimeout = function() {
-            debug('_transportTimeout');
-            if (this.readyState === SockJS.CONNECTING) {
-              this._transportClose(2007, 'Transport timed out');
-            }
-          };
-
-          SockJS.prototype._transportMessage = function(msg) {
-            debug('_transportMessage', msg);
-            var self = this,
-              type = msg.slice(0, 1),
-              content = msg.slice(1),
-              payload;
-
-            // first check for messages that don't need a payload
-            switch (type) {
-              case 'o':
-                this._open();
-                return;
-              case 'h':
-                this.dispatchEvent(new Event('heartbeat'));
-                debug('heartbeat', this.transport);
-                return;
-            }
-
-            if (content) {
-              try {
-                payload = JSON3.parse(content);
-              } catch (e) {
-                debug('bad json', content);
-              }
-            }
-
-            if (typeof payload === 'undefined') {
-              debug('empty payload', content);
-              return;
-            }
-
-            switch (type) {
-              case 'a':
-                if (Array.isArray(payload)) {
-                  payload.forEach(function(p) {
-                    debug('message', self.transport, p);
-                    self.dispatchEvent(new TransportMessageEvent(p));
-                  });
-                }
-                break;
-              case 'm':
-                debug('message', this.transport, payload);
-                this.dispatchEvent(new TransportMessageEvent(payload));
-                break;
-              case 'c':
-                if (Array.isArray(payload) && payload.length === 2) {
-                  this._close(payload[0], payload[1], true);
-                }
-                break;
-            }
-          };
-
-          SockJS.prototype._transportClose = function(code, reason) {
-            debug('_transportClose', this.transport, code, reason);
-            if (this._transport) {
-              this._transport.removeAllListeners();
-              this._transport = null;
-              this.transport = null;
-            }
-
-            if (!userSetCode(code) && code !== 2000 && this.readyState === SockJS.CONNECTING) {
-              this._connect();
-              return;
-            }
-
-            this._close(code, reason);
-          };
-
-          SockJS.prototype._open = function() {
-            debug('_open', this._transport.transportName, this.readyState);
-            if (this.readyState === SockJS.CONNECTING) {
-              if (this._transportTimeoutId) {
-                clearTimeout(this._transportTimeoutId);
-                this._transportTimeoutId = null;
-              }
-              this.readyState = SockJS.OPEN;
-              this.transport = this._transport.transportName;
-              this.dispatchEvent(new Event('open'));
-              debug('connected', this.transport);
-            } else {
-              // The server might have been restarted, and lost track of our
-              // connection.
-              this._close(1006, 'Server lost session');
-            }
-          };
-
-          SockJS.prototype._close = function(code, reason, wasClean) {
-            debug('_close', this.transport, code, reason, wasClean, this.readyState);
-            var forceFail = false;
-
-            if (this._ir) {
-              forceFail = true;
-              this._ir.close();
-              this._ir = null;
-            }
-            if (this._transport) {
-              this._transport.close();
-              this._transport = null;
-              this.transport = null;
-            }
-
-            if (this.readyState === SockJS.CLOSED) {
-              throw new Error('InvalidStateError: SockJS has already been closed');
-            }
-
-            this.readyState = SockJS.CLOSING;
-            setTimeout(function() {
-              this.readyState = SockJS.CLOSED;
-
-              if (forceFail) {
-                this.dispatchEvent(new Event('error'));
-              }
-
-              var e = new CloseEvent('close');
-              e.wasClean = wasClean || false;
-              e.code = code || 1000;
-              e.reason = reason;
-
-              this.dispatchEvent(e);
-              this.onmessage = this.onclose = this.onerror = null;
-              debug('disconnected');
-            }.bind(this), 0);
-          };
-
-          // See: http://www.erg.abdn.ac.uk/~gerrit/dccp/notes/ccid2/rto_estimator/
-          // and RFC 2988.
-          SockJS.prototype.countRTO = function(rtt) {
-            // In a local environment, when using IE8/9 and the `jsonp-polling`
-            // transport the time needed to establish a connection (the time that pass
-            // from the opening of the transport to the call of `_dispatchOpen`) is
-            // around 200msec (the lower bound used in the article above) and this
-            // causes spurious timeouts. For this reason we calculate a value slightly
-            // larger than that used in the article.
-            if (rtt > 100) {
-              return 4 * rtt; // rto > 400msec
-            }
-            return 300 + rtt; // 300msec < rto <= 400msec
-          };
-
-          module.exports = function(availableTransports) {
-            transports = transport(availableTransports);
-            require('./iframe-bootstrap')(SockJS, availableTransports);
-            return SockJS;
-          };
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./event/close": 2,
-        "./event/event": 4,
-        "./event/eventtarget": 5,
-        "./event/trans-message": 6,
-        "./iframe-bootstrap": 8,
-        "./info-receiver": 12,
-        "./location": 13,
-        "./shims": 15,
-        "./utils/browser": 44,
-        "./utils/escape": 45,
-        "./utils/event": 46,
-        "./utils/log": 48,
-        "./utils/object": 49,
-        "./utils/random": 50,
-        "./utils/transport": 51,
-        "./utils/url": 52,
-        "./version": 53,
-        "debug": 55,
-        "inherits": 57,
-        "json3": 58,
-        "url-parse": 61
-      }],
-      15: [function(require, module, exports) {
-        /* eslint-disable */
-        /* jscs: disable */
-        'use strict';
-
-        // pulled specific shims from https://github.com/es-shims/es5-shim
-
-        var ArrayPrototype = Array.prototype;
-        var ObjectPrototype = Object.prototype;
-        var FunctionPrototype = Function.prototype;
-        var StringPrototype = String.prototype;
-        var array_slice = ArrayPrototype.slice;
-
-        var _toString = ObjectPrototype.toString;
-        var isFunction = function(val) {
-          return ObjectPrototype.toString.call(val) === '[object Function]';
-        };
-        var isArray = function isArray(obj) {
-          return _toString.call(obj) === '[object Array]';
-        };
-        var isString = function isString(obj) {
-          return _toString.call(obj) === '[object String]';
-        };
-
-        var supportsDescriptors = Object.defineProperty && (function() {
-          try {
-            Object.defineProperty({}, 'x', {});
-            return true;
-          } catch (e) { /* this is ES3 */
-            return false;
-          }
-        }());
-
-        // Define configurable, writable and non-enumerable props
-        // if they don't exist.
-        var defineProperty;
-        if (supportsDescriptors) {
-          defineProperty = function(object, name, method, forceAssign) {
-            if (!forceAssign && (name in object)) {
-              return;
-            }
-            Object.defineProperty(object, name, {
-              configurable: true,
-              enumerable: false,
-              writable: true,
-              value: method
-            });
-          };
-        } else {
-          defineProperty = function(object, name, method, forceAssign) {
-            if (!forceAssign && (name in object)) {
-              return;
-            }
-            object[name] = method;
-          };
-        }
-        var defineProperties = function(object, map, forceAssign) {
-          for (var name in map) {
-            if (ObjectPrototype.hasOwnProperty.call(map, name)) {
-              defineProperty(object, name, map[name], forceAssign);
-            }
-          }
-        };
-
-        var toObject = function(o) {
-          if (o == null) { // this matches both null and undefined
-            throw new TypeError("can't convert " + o + ' to object');
-          }
-          return Object(o);
-        };
-
-        //
-        // Util
-        // ======
-        //
-
-        // ES5 9.4
-        // http://es5.github.com/#x9.4
-        // http://jsperf.com/to-integer
-
-        function toInteger(num) {
-          var n = +num;
-          if (n !== n) { // isNaN
-            n = 0;
-          } else if (n !== 0 && n !== (1 / 0) && n !== -(1 / 0)) {
-            n = (n > 0 || -1) * Math.floor(Math.abs(n));
-          }
-          return n;
-        }
-
-        function ToUint32(x) {
-          return x >>> 0;
-        }
-
-        //
-        // Function
-        // ========
-        //
-
-        // ES-5 15.3.4.5
-        // http://es5.github.com/#x15.3.4.5
-
-        function Empty() {}
-
-        defineProperties(FunctionPrototype, {
-          bind: function bind(that) { // .length is 1
-            // 1. Let Target be the this value.
-            var target = this;
-            // 2. If IsCallable(Target) is false, throw a TypeError exception.
-            if (!isFunction(target)) {
-              throw new TypeError('Function.prototype.bind called on incompatible ' + target);
-            }
-            // 3. Let A be a new (possibly empty) internal list of all of the
-            //   argument values provided after thisArg (arg1, arg2 etc), in order.
-            // XXX slicedArgs will stand in for "A" if used
-            var args = array_slice.call(arguments, 1); // for normal call
-            // 4. Let F be a new native ECMAScript object.
-            // 11. Set the [[Prototype]] internal property of F to the standard
-            //   built-in Function prototype object as specified in 15.3.3.1.
-            // 12. Set the [[Call]] internal property of F as described in
-            //   15.3.4.5.1.
-            // 13. Set the [[Construct]] internal property of F as described in
-            //   15.3.4.5.2.
-            // 14. Set the [[HasInstance]] internal property of F as described in
-            //   15.3.4.5.3.
-            var binder = function() {
-
-              if (this instanceof bound) {
-                // 15.3.4.5.2 [[Construct]]
-                // When the [[Construct]] internal method of a function object,
-                // F that was created using the bind function is called with a
-                // list of arguments ExtraArgs, the following steps are taken:
-                // 1. Let target be the value of F's [[TargetFunction]]
-                //   internal property.
-                // 2. If target has no [[Construct]] internal method, a
-                //   TypeError exception is thrown.
-                // 3. Let boundArgs be the value of F's [[BoundArgs]] internal
-                //   property.
-                // 4. Let args be a new list containing the same values as the
-                //   list boundArgs in the same order followed by the same
-                //   values as the list ExtraArgs in the same order.
-                // 5. Return the result of calling the [[Construct]] internal
-                //   method of target providing args as the arguments.
-
-                var result = target.apply(
-                  this,
-                  args.concat(array_slice.call(arguments))
-                );
-                if (Object(result) === result) {
-                  return result;
-                }
-                return this;
-
-              } else {
-                // 15.3.4.5.1 [[Call]]
-                // When the [[Call]] internal method of a function object, F,
-                // which was created using the bind function is called with a
-                // this value and a list of arguments ExtraArgs, the following
-                // steps are taken:
-                // 1. Let boundArgs be the value of F's [[BoundArgs]] internal
-                //   property.
-                // 2. Let boundThis be the value of F's [[BoundThis]] internal
-                //   property.
-                // 3. Let target be the value of F's [[TargetFunction]] internal
-                //   property.
-                // 4. Let args be a new list containing the same values as the
-                //   list boundArgs in the same order followed by the same
-                //   values as the list ExtraArgs in the same order.
-                // 5. Return the result of calling the [[Call]] internal method
-                //   of target providing boundThis as the this value and
-                //   providing args as the arguments.
-
-                // equiv: target.call(this, ...boundArgs, ...args)
-                return target.apply(
-                  that,
-                  args.concat(array_slice.call(arguments))
-                );
-
-              }
-
-            };
-
-            // 15. If the [[Class]] internal property of Target is "Function", then
-            //     a. Let L be the length property of Target minus the length of A.
-            //     b. Set the length own property of F to either 0 or L, whichever is
-            //       larger.
-            // 16. Else set the length own property of F to 0.
-
-            var boundLength = Math.max(0, target.length - args.length);
-
-            // 17. Set the attributes of the length own property of F to the values
-            //   specified in 15.3.5.1.
-            var boundArgs = [];
-            for (var i = 0; i < boundLength; i++) {
-              boundArgs.push('$' + i);
-            }
-
-            // XXX Build a dynamic function with desired amount of arguments is the only
-            // way to set the length property of a function.
-            // In environments where Content Security Policies enabled (Chrome extensions,
-            // for ex.) all use of eval or Function costructor throws an exception.
-            // However in all of these environments Function.prototype.bind exists
-            // and so this code will never be executed.
-            var bound = Function('binder', 'return function (' + boundArgs.join(',') + '){ return binder.apply(this, arguments); }')(binder);
-
-            if (target.prototype) {
-              Empty.prototype = target.prototype;
-              bound.prototype = new Empty();
-              // Clean up dangling references.
-              Empty.prototype = null;
-            }
-
-            // TODO
-            // 18. Set the [[Extensible]] internal property of F to true.
-
-            // TODO
-            // 19. Let thrower be the [[ThrowTypeError]] function Object (13.2.3).
-            // 20. Call the [[DefineOwnProperty]] internal method of F with
-            //   arguments "caller", PropertyDescriptor {[[Get]]: thrower, [[Set]]:
-            //   thrower, [[Enumerable]]: false, [[Configurable]]: false}, and
-            //   false.
-            // 21. Call the [[DefineOwnProperty]] internal method of F with
-            //   arguments "arguments", PropertyDescriptor {[[Get]]: thrower,
-            //   [[Set]]: thrower, [[Enumerable]]: false, [[Configurable]]: false},
-            //   and false.
-
-            // TODO
-            // NOTE Function objects created using Function.prototype.bind do not
-            // have a prototype property or the [[Code]], [[FormalParameters]], and
-            // [[Scope]] internal properties.
-            // XXX can't delete prototype in pure-js.
-
-            // 22. Return F.
-            return bound;
-          }
-        });
-
-        //
-        // Array
-        // =====
-        //
-
-        // ES5 15.4.3.2
-        // http://es5.github.com/#x15.4.3.2
-        // https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/Array/isArray
-        defineProperties(Array, {
-          isArray: isArray
-        });
-
-
-        var boxedString = Object('a');
-        var splitString = boxedString[0] !== 'a' || !(0 in boxedString);
-
-        var properlyBoxesContext = function properlyBoxed(method) {
-          // Check node 0.6.21 bug where third parameter is not boxed
-          var properlyBoxesNonStrict = true;
-          var properlyBoxesStrict = true;
-          if (method) {
-            method.call('foo', function(_, __, context) {
-              if (typeof context !== 'object') {
-                properlyBoxesNonStrict = false;
-              }
-            });
-
-            method.call([1], function() {
-              'use strict';
-              properlyBoxesStrict = typeof this === 'string';
-            }, 'x');
-          }
-          return !!method && properlyBoxesNonStrict && properlyBoxesStrict;
-        };
-
-        defineProperties(ArrayPrototype, {
-          forEach: function forEach(fun /*, thisp*/ ) {
-            var object = toObject(this),
-              self = splitString && isString(this) ? this.split('') : object,
-              thisp = arguments[1],
-              i = -1,
-              length = self.length >>> 0;
-
-            // If no callback function or if callback is not a callable function
-            if (!isFunction(fun)) {
-              throw new TypeError(); // TODO message
-            }
-
-            while (++i < length) {
-              if (i in self) {
-                // Invoke the callback function with call, passing arguments:
-                // context, property value, property key, thisArg object
-                // context
-                fun.call(thisp, self[i], i, object);
-              }
-            }
-          }
-        }, !properlyBoxesContext(ArrayPrototype.forEach));
-
-        // ES5 15.4.4.14
-        // http://es5.github.com/#x15.4.4.14
-        // https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/Array/indexOf
-        var hasFirefox2IndexOfBug = Array.prototype.indexOf && [0, 1].indexOf(1, 2) !== -1;
-        defineProperties(ArrayPrototype, {
-          indexOf: function indexOf(sought /*, fromIndex */ ) {
-            var self = splitString && isString(this) ? this.split('') : toObject(this),
-              length = self.length >>> 0;
-
-            if (!length) {
-              return -1;
-            }
-
-            var i = 0;
-            if (arguments.length > 1) {
-              i = toInteger(arguments[1]);
-            }
-
-            // handle negative indices
-            i = i >= 0 ? i : Math.max(0, length + i);
-            for (; i < length; i++) {
-              if (i in self && self[i] === sought) {
-                return i;
-              }
-            }
-            return -1;
-          }
-        }, hasFirefox2IndexOfBug);
-
-        //
-        // String
-        // ======
-        //
-
-        // ES5 15.5.4.14
-        // http://es5.github.com/#x15.5.4.14
-
-        // [bugfix, IE lt 9, firefox 4, Konqueror, Opera, obscure browsers]
-        // Many browsers do not split properly with regular expressions or they
-        // do not perform the split correctly under obscure conditions.
-        // See http://blog.stevenlevithan.com/archives/cross-browser-split
-        // I've tested in many browsers and this seems to cover the deviant ones:
-        //    'ab'.split(/(?:ab)*/) should be ["", ""], not [""]
-        //    '.'.split(/(.?)(.?)/) should be ["", ".", "", ""], not ["", ""]
-        //    'tesst'.split(/(s)*/) should be ["t", undefined, "e", "s", "t"], not
-        //       [undefined, "t", undefined, "e", ...]
-        //    ''.split(/.?/) should be [], not [""]
-        //    '.'.split(/()()/) should be ["."], not ["", "", "."]
-
-        var string_split = StringPrototype.split;
-        if (
-          'ab'.split(/(?:ab)*/).length !== 2 ||
-          '.'.split(/(.?)(.?)/).length !== 4 ||
-          'tesst'.split(/(s)*/)[1] === 't' ||
-          'test'.split(/(?:)/, -1).length !== 4 ||
-          ''.split(/.?/).length ||
-          '.'.split(/()()/).length > 1
-        ) {
-          (function() {
-            var compliantExecNpcg = /()??/.exec('')[1] === void 0; // NPCG: nonparticipating capturing group
-
-            StringPrototype.split = function(separator, limit) {
-              var string = this;
-              if (separator === void 0 && limit === 0) {
-                return [];
-              }
-
-              // If `separator` is not a regex, use native split
-              if (_toString.call(separator) !== '[object RegExp]') {
-                return string_split.call(this, separator, limit);
-              }
-
-              var output = [],
-                flags = (separator.ignoreCase ? 'i' : '') +
-                (separator.multiline ? 'm' : '') +
-                (separator.extended ? 'x' : '') + // Proposed for ES6
-                (separator.sticky ? 'y' : ''), // Firefox 3+
-                lastLastIndex = 0,
-                // Make `global` and avoid `lastIndex` issues by working with a copy
-                separator2, match, lastIndex, lastLength;
-              separator = new RegExp(separator.source, flags + 'g');
-              string += ''; // Type-convert
-              if (!compliantExecNpcg) {
-                // Doesn't need flags gy, but they don't hurt
-                separator2 = new RegExp('^' + separator.source + '$(?!\\s)', flags);
-              }
-              /* Values for `limit`, per the spec:
-               * If undefined: 4294967295 // Math.pow(2, 32) - 1
-               * If 0, Infinity, or NaN: 0
-               * If positive number: limit = Math.floor(limit); if (limit > 4294967295) limit -= 4294967296;
-               * If negative number: 4294967296 - Math.floor(Math.abs(limit))
-               * If other: Type-convert, then use the above rules
-               */
-              limit = limit === void 0 ?
-                -1 >>> 0 : // Math.pow(2, 32) - 1
-                ToUint32(limit);
-              while (match = separator.exec(string)) {
-                // `separator.lastIndex` is not reliable cross-browser
-                lastIndex = match.index + match[0].length;
-                if (lastIndex > lastLastIndex) {
-                  output.push(string.slice(lastLastIndex, match.index));
-                  // Fix browsers whose `exec` methods don't consistently return `undefined` for
-                  // nonparticipating capturing groups
-                  if (!compliantExecNpcg && match.length > 1) {
-                    match[0].replace(separator2, function() {
-                      for (var i = 1; i < arguments.length - 2; i++) {
-                        if (arguments[i] === void 0) {
-                          match[i] = void 0;
-                        }
-                      }
-                    });
-                  }
-                  if (match.length > 1 && match.index < string.length) {
-                    ArrayPrototype.push.apply(output, match.slice(1));
-                  }
-                  lastLength = match[0].length;
-                  lastLastIndex = lastIndex;
-                  if (output.length >= limit) {
-                    break;
-                  }
-                }
-                if (separator.lastIndex === match.index) {
-                  separator.lastIndex++; // Avoid an infinite loop
-                }
-              }
-              if (lastLastIndex === string.length) {
-                if (lastLength || !separator.test('')) {
-                  output.push('');
-                }
-              } else {
-                output.push(string.slice(lastLastIndex));
-              }
-              return output.length > limit ? output.slice(0, limit) : output;
-            };
-          }());
-
-          // [bugfix, chrome]
-          // If separator is undefined, then the result array contains just one String,
-          // which is the this value (converted to a String). If limit is not undefined,
-          // then the output array is truncated so that it contains no more than limit
-          // elements.
-          // "0".split(undefined, 0) -> []
-        } else if ('0'.split(void 0, 0).length) {
-          StringPrototype.split = function split(separator, limit) {
-            if (separator === void 0 && limit === 0) {
-              return [];
-            }
-            return string_split.call(this, separator, limit);
-          };
-        }
-
-        // ECMA-262, 3rd B.2.3
-        // Not an ECMAScript standard, although ECMAScript 3rd Edition has a
-        // non-normative section suggesting uniform semantics and it should be
-        // normalized across all browsers
-        // [bugfix, IE lt 9] IE < 9 substr() with negative value not working in IE
-        var string_substr = StringPrototype.substr;
-        var hasNegativeSubstrBug = ''.substr && '0b'.substr(-1) !== 'b';
-        defineProperties(StringPrototype, {
-          substr: function substr(start, length) {
-            return string_substr.call(
-              this,
-              start < 0 ? ((start = this.length + start) < 0 ? 0 : start) : start,
-              length
-            );
-          }
-        }, hasNegativeSubstrBug);
-
-      }, {}],
-      16: [function(require, module, exports) {
-        'use strict';
-
-        module.exports = [
-          // streaming transports
-          require('./transport/websocket'), require('./transport/xhr-streaming'), require('./transport/xdr-streaming'), require('./transport/eventsource'), require('./transport/lib/iframe-wrap')(require('./transport/eventsource'))
-
-          // polling transports
-          , require('./transport/htmlfile'), require('./transport/lib/iframe-wrap')(require('./transport/htmlfile')), require('./transport/xhr-polling'), require('./transport/xdr-polling'), require('./transport/lib/iframe-wrap')(require('./transport/xhr-polling')), require('./transport/jsonp-polling')
-        ];
-
-      }, {
-        "./transport/eventsource": 20,
-        "./transport/htmlfile": 21,
-        "./transport/jsonp-polling": 23,
-        "./transport/lib/iframe-wrap": 26,
-        "./transport/websocket": 38,
-        "./transport/xdr-polling": 39,
-        "./transport/xdr-streaming": 40,
-        "./transport/xhr-polling": 41,
-        "./transport/xhr-streaming": 42
-      }],
-      17: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var EventEmitter = require('events').EventEmitter,
-            inherits = require('inherits'),
-            utils = require('../../utils/event'),
-            urlUtils = require('../../utils/url'),
-            XHR = global.XMLHttpRequest;
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:browser:xhr');
-          }
-
-          function AbstractXHRObject(method, url, payload, opts) {
-            debug(method, url);
-            var self = this;
-            EventEmitter.call(this);
-
-            setTimeout(function() {
-              self._start(method, url, payload, opts);
-            }, 0);
-          }
-
-          inherits(AbstractXHRObject, EventEmitter);
-
-          AbstractXHRObject.prototype._start = function(method, url, payload, opts) {
-            var self = this;
-
-            try {
-              this.xhr = new XHR();
-            } catch (x) {
-              // intentionally empty
-            }
-
-            if (!this.xhr) {
-              debug('no xhr');
-              this.emit('finish', 0, 'no xhr support');
-              this._cleanup();
-              return;
-            }
-
-            // several browsers cache POSTs
-            url = urlUtils.addQuery(url, 't=' + (+new Date()));
-
-            // Explorer tends to keep connection open, even after the
-            // tab gets closed: http://bugs.jquery.com/ticket/5280
-            this.unloadRef = utils.unloadAdd(function() {
-              debug('unload cleanup');
-              self._cleanup(true);
-            });
-            try {
-              this.xhr.open(method, url, true);
-              if (this.timeout && 'timeout' in this.xhr) {
-                this.xhr.timeout = this.timeout;
-                this.xhr.ontimeout = function() {
-                  debug('xhr timeout');
-                  self.emit('finish', 0, '');
-                  self._cleanup(false);
-                };
-              }
-            } catch (e) {
-              debug('exception', e);
-              // IE raises an exception on wrong port.
-              this.emit('finish', 0, '');
-              this._cleanup(false);
-              return;
-            }
-
-            if ((!opts || !opts.noCredentials) && AbstractXHRObject.supportsCORS) {
-              debug('withCredentials');
-              // Mozilla docs says https://developer.mozilla.org/en/XMLHttpRequest :
-              // "This never affects same-site requests."
-
-              this.xhr.withCredentials = 'true';
-            }
-            if (opts && opts.headers) {
-              for (var key in opts.headers) {
-                this.xhr.setRequestHeader(key, opts.headers[key]);
-              }
-            }
-
-            this.xhr.onreadystatechange = function() {
-              if (self.xhr) {
-                var x = self.xhr;
-                var text, status;
-                debug('readyState', x.readyState);
-                switch (x.readyState) {
-                  case 3:
-                    // IE doesn't like peeking into responseText or status
-                    // on Microsoft.XMLHTTP and readystate=3
-                    try {
-                      status = x.status;
-                      text = x.responseText;
-                    } catch (e) {
-                      // intentionally empty
-                    }
-                    debug('status', status);
-                    // IE returns 1223 for 204: http://bugs.jquery.com/ticket/1450
-                    if (status === 1223) {
-                      status = 204;
-                    }
-
-                    // IE does return readystate == 3 for 404 answers.
-                    if (status === 200 && text && text.length > 0) {
-                      debug('chunk');
-                      self.emit('chunk', status, text);
-                    }
-                    break;
-                  case 4:
-                    status = x.status;
-                    debug('status', status);
-                    // IE returns 1223 for 204: http://bugs.jquery.com/ticket/1450
-                    if (status === 1223) {
-                      status = 204;
-                    }
-                    // IE returns this for a bad port
-                    // http://msdn.microsoft.com/en-us/library/windows/desktop/aa383770(v=vs.85).aspx
-                    if (status === 12005 || status === 12029) {
-                      status = 0;
-                    }
-
-                    debug('finish', status, x.responseText);
-                    self.emit('finish', status, x.responseText);
-                    self._cleanup(false);
-                    break;
-                }
-              }
-            };
-
-            try {
-              self.xhr.send(payload);
-            } catch (e) {
-              self.emit('finish', 0, '');
-              self._cleanup(false);
-            }
-          };
-
-          AbstractXHRObject.prototype._cleanup = function(abort) {
-            debug('cleanup');
-            if (!this.xhr) {
-              return;
-            }
-            this.removeAllListeners();
-            utils.unloadDel(this.unloadRef);
-
-            // IE needs this field to be a function
-            this.xhr.onreadystatechange = function() {};
-            if (this.xhr.ontimeout) {
-              this.xhr.ontimeout = null;
-            }
-
-            if (abort) {
-              try {
-                this.xhr.abort();
-              } catch (x) {
-                // intentionally empty
-              }
-            }
-            this.unloadRef = this.xhr = null;
-          };
-
-          AbstractXHRObject.prototype.close = function() {
-            debug('close');
-            this._cleanup(true);
-          };
-
-          AbstractXHRObject.enabled = !!XHR;
-          // override XMLHttpRequest for IE6/7
-          // obfuscate to avoid firewalls
-          var axo = ['Active'].concat('Object').join('X');
-          if (!AbstractXHRObject.enabled && (axo in global)) {
-            debug('overriding xmlhttprequest');
-            XHR = function() {
-              try {
-                return new global[axo]('Microsoft.XMLHTTP');
-              } catch (e) {
-                return null;
-              }
-            };
-            AbstractXHRObject.enabled = !!new XHR();
-          }
-
-          var cors = false;
-          try {
-            cors = 'withCredentials' in new XHR();
-          } catch (ignored) {
-            // intentionally empty
-          }
-
-          AbstractXHRObject.supportsCORS = cors;
-
-          module.exports = AbstractXHRObject;
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../../utils/event": 46,
-        "../../utils/url": 52,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      18: [function(require, module, exports) {
-        (function(global) {
-          module.exports = global.EventSource;
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      19: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var Driver = global.WebSocket || global.MozWebSocket;
-          if (Driver) {
-            module.exports = function WebSocketBrowserDriver(url) {
-              return new Driver(url);
-            };
-          } else {
-            module.exports = undefined;
-          }
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      20: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          AjaxBasedTransport = require('./lib/ajax-based'),
-          EventSourceReceiver = require('./receiver/eventsource'),
-          XHRCorsObject = require('./sender/xhr-cors'),
-          EventSourceDriver = require('eventsource');
-
-        function EventSourceTransport(transUrl) {
-          if (!EventSourceTransport.enabled()) {
-            throw new Error('Transport created when disabled');
-          }
-
-          AjaxBasedTransport.call(this, transUrl, '/eventsource', EventSourceReceiver, XHRCorsObject);
-        }
-
-        inherits(EventSourceTransport, AjaxBasedTransport);
-
-        EventSourceTransport.enabled = function() {
-          return !!EventSourceDriver;
-        };
-
-        EventSourceTransport.transportName = 'eventsource';
-        EventSourceTransport.roundTrips = 2;
-
-        module.exports = EventSourceTransport;
-
-      }, {
-        "./lib/ajax-based": 24,
-        "./receiver/eventsource": 29,
-        "./sender/xhr-cors": 35,
-        "eventsource": 18,
-        "inherits": 57
-      }],
-      21: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          HtmlfileReceiver = require('./receiver/htmlfile'),
-          XHRLocalObject = require('./sender/xhr-local'),
-          AjaxBasedTransport = require('./lib/ajax-based');
-
-        function HtmlFileTransport(transUrl) {
-          if (!HtmlfileReceiver.enabled) {
-            throw new Error('Transport created when disabled');
-          }
-          AjaxBasedTransport.call(this, transUrl, '/htmlfile', HtmlfileReceiver, XHRLocalObject);
-        }
-
-        inherits(HtmlFileTransport, AjaxBasedTransport);
-
-        HtmlFileTransport.enabled = function(info) {
-          return HtmlfileReceiver.enabled && info.sameOrigin;
-        };
-
-        HtmlFileTransport.transportName = 'htmlfile';
-        HtmlFileTransport.roundTrips = 2;
-
-        module.exports = HtmlFileTransport;
-
-      }, {
-        "./lib/ajax-based": 24,
-        "./receiver/htmlfile": 30,
-        "./sender/xhr-local": 37,
-        "inherits": 57
-      }],
-      22: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          // Few cool transports do work only for same-origin. In order to make
-          // them work cross-domain we shall use iframe, served from the
-          // remote domain. New browsers have capabilities to communicate with
-          // cross domain iframe using postMessage(). In IE it was implemented
-          // from IE 8+, but of course, IE got some details wrong:
-          //    http://msdn.microsoft.com/en-us/library/cc197015(v=VS.85).aspx
-          //    http://stevesouders.com/misc/test-postmessage.php
-
-          var inherits = require('inherits'),
-            JSON3 = require('json3'),
-            EventEmitter = require('events').EventEmitter,
-            version = require('../version'),
-            urlUtils = require('../utils/url'),
-            iframeUtils = require('../utils/iframe'),
-            eventUtils = require('../utils/event'),
-            random = require('../utils/random');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:transport:iframe');
-          }
-
-          function IframeTransport(transport, transUrl, baseUrl) {
-            if (!IframeTransport.enabled()) {
-              throw new Error('Transport created when disabled');
-            }
-            EventEmitter.call(this);
-
-            var self = this;
-            this.origin = urlUtils.getOrigin(baseUrl);
-            this.baseUrl = baseUrl;
-            this.transUrl = transUrl;
-            this.transport = transport;
-            this.windowId = random.string(8);
-
-            var iframeUrl = urlUtils.addPath(baseUrl, '/iframe.html') + '#' + this.windowId;
-            debug(transport, transUrl, iframeUrl);
-
-            this.iframeObj = iframeUtils.createIframe(iframeUrl, function(r) {
-              debug('err callback');
-              self.emit('close', 1006, 'Unable to load an iframe (' + r + ')');
-              self.close();
-            });
-
-            this.onmessageCallback = this._message.bind(this);
-            eventUtils.attachEvent('message', this.onmessageCallback);
-          }
-
-          inherits(IframeTransport, EventEmitter);
-
-          IframeTransport.prototype.close = function() {
-            debug('close');
-            this.removeAllListeners();
-            if (this.iframeObj) {
-              eventUtils.detachEvent('message', this.onmessageCallback);
-              try {
-                // When the iframe is not loaded, IE raises an exception
-                // on 'contentWindow'.
-                this.postMessage('c');
-              } catch (x) {
-                // intentionally empty
-              }
-              this.iframeObj.cleanup();
-              this.iframeObj = null;
-              this.onmessageCallback = this.iframeObj = null;
-            }
-          };
-
-          IframeTransport.prototype._message = function(e) {
-            debug('message', e.data);
-            if (!urlUtils.isOriginEqual(e.origin, this.origin)) {
-              debug('not same origin', e.origin, this.origin);
-              return;
-            }
-
-            var iframeMessage;
-            try {
-              iframeMessage = JSON3.parse(e.data);
-            } catch (ignored) {
-              debug('bad json', e.data);
-              return;
-            }
-
-            if (iframeMessage.windowId !== this.windowId) {
-              debug('mismatched window id', iframeMessage.windowId, this.windowId);
-              return;
-            }
-
-            switch (iframeMessage.type) {
-              case 's':
-                this.iframeObj.loaded();
-                // window global dependency
-                this.postMessage('s', JSON3.stringify([
-                  version, this.transport, this.transUrl, this.baseUrl
-                ]));
-                break;
-              case 't':
-                this.emit('message', iframeMessage.data);
-                break;
-              case 'c':
-                var cdata;
-                try {
-                  cdata = JSON3.parse(iframeMessage.data);
-                } catch (ignored) {
-                  debug('bad json', iframeMessage.data);
-                  return;
-                }
-                this.emit('close', cdata[0], cdata[1]);
-                this.close();
-                break;
-            }
-          };
-
-          IframeTransport.prototype.postMessage = function(type, data) {
-            debug('postMessage', type, data);
-            this.iframeObj.post(JSON3.stringify({
-              windowId: this.windowId,
-              type: type,
-              data: data || ''
-            }), this.origin);
-          };
-
-          IframeTransport.prototype.send = function(message) {
-            debug('send', message);
-            this.postMessage('m', message);
-          };
-
-          IframeTransport.enabled = function() {
-            return iframeUtils.iframeEnabled;
-          };
-
-          IframeTransport.transportName = 'iframe';
-          IframeTransport.roundTrips = 2;
-
-          module.exports = IframeTransport;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "../utils/event": 46,
-        "../utils/iframe": 47,
-        "../utils/random": 50,
-        "../utils/url": 52,
-        "../version": 53,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57,
-        "json3": 58
-      }],
-      23: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          // The simplest and most robust transport, using the well-know cross
-          // domain hack - JSONP. This transport is quite inefficient - one
-          // message could use up to one http request. But at least it works almost
-          // everywhere.
-          // Known limitations:
-          //   o you will get a spinning cursor
-          //   o for Konqueror a dumb timer is needed to detect errors
-
-          var inherits = require('inherits'),
-            SenderReceiver = require('./lib/sender-receiver'),
-            JsonpReceiver = require('./receiver/jsonp'),
-            jsonpSender = require('./sender/jsonp');
-
-          function JsonPTransport(transUrl) {
-            if (!JsonPTransport.enabled()) {
-              throw new Error('Transport created when disabled');
-            }
-            SenderReceiver.call(this, transUrl, '/jsonp', jsonpSender, JsonpReceiver);
-          }
-
-          inherits(JsonPTransport, SenderReceiver);
-
-          JsonPTransport.enabled = function() {
-            return !!global.document;
-          };
-
-          JsonPTransport.transportName = 'jsonp-polling';
-          JsonPTransport.roundTrips = 1;
-          JsonPTransport.needBody = true;
-
-          module.exports = JsonPTransport;
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./lib/sender-receiver": 28,
-        "./receiver/jsonp": 31,
-        "./sender/jsonp": 33,
-        "inherits": 57
-      }],
-      24: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            urlUtils = require('../../utils/url'),
-            SenderReceiver = require('./sender-receiver');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:ajax-based');
-          }
-
-          function createAjaxSender(AjaxObject) {
-            return function(url, payload, callback) {
-              debug('create ajax sender', url, payload);
-              var opt = {};
-              if (typeof payload === 'string') {
-                opt.headers = {
-                  'Content-type': 'text/plain'
-                };
-              }
-              var ajaxUrl = urlUtils.addPath(url, '/xhr_send');
-              var xo = new AjaxObject('POST', ajaxUrl, payload, opt);
-              xo.once('finish', function(status) {
-                debug('finish', status);
-                xo = null;
-
-                if (status !== 200 && status !== 204) {
-                  return callback(new Error('http status ' + status));
-                }
-                callback();
-              });
-              return function() {
-                debug('abort');
-                xo.close();
-                xo = null;
-
-                var err = new Error('Aborted');
-                err.code = 1000;
-                callback(err);
-              };
-            };
-          }
-
-          function AjaxBasedTransport(transUrl, urlSuffix, Receiver, AjaxObject) {
-            SenderReceiver.call(this, transUrl, urlSuffix, createAjaxSender(AjaxObject), Receiver, AjaxObject);
-          }
-
-          inherits(AjaxBasedTransport, SenderReceiver);
-
-          module.exports = AjaxBasedTransport;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "../../utils/url": 52,
-        "./sender-receiver": 28,
-        "debug": 55,
-        "inherits": 57
-      }],
-      25: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            EventEmitter = require('events').EventEmitter;
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:buffered-sender');
-          }
-
-          function BufferedSender(url, sender) {
-            debug(url);
-            EventEmitter.call(this);
-            this.sendBuffer = [];
-            this.sender = sender;
-            this.url = url;
-          }
-
-          inherits(BufferedSender, EventEmitter);
-
-          BufferedSender.prototype.send = function(message) {
-            debug('send', message);
-            this.sendBuffer.push(message);
-            if (!this.sendStop) {
-              this.sendSchedule();
-            }
-          };
-
-          // For polling transports in a situation when in the message callback,
-          // new message is being send. If the sending connection was started
-          // before receiving one, it is possible to saturate the network and
-          // timeout due to the lack of receiving socket. To avoid that we delay
-          // sending messages by some small time, in order to let receiving
-          // connection be started beforehand. This is only a halfmeasure and
-          // does not fix the big problem, but it does make the tests go more
-          // stable on slow networks.
-          BufferedSender.prototype.sendScheduleWait = function() {
-            debug('sendScheduleWait');
-            var self = this;
-            var tref;
-            this.sendStop = function() {
-              debug('sendStop');
-              self.sendStop = null;
-              clearTimeout(tref);
-            };
-            tref = setTimeout(function() {
-              debug('timeout');
-              self.sendStop = null;
-              self.sendSchedule();
-            }, 25);
-          };
-
-          BufferedSender.prototype.sendSchedule = function() {
-            debug('sendSchedule', this.sendBuffer.length);
-            var self = this;
-            if (this.sendBuffer.length > 0) {
-              var payload = '[' + this.sendBuffer.join(',') + ']';
-              this.sendStop = this.sender(this.url, payload, function(err) {
-                self.sendStop = null;
-                if (err) {
-                  debug('error', err);
-                  self.emit('close', err.code || 1006, 'Sending error: ' + err);
-                  self.close();
-                } else {
-                  self.sendScheduleWait();
-                }
-              });
-              this.sendBuffer = [];
-            }
-          };
-
-          BufferedSender.prototype._cleanup = function() {
-            debug('_cleanup');
-            this.removeAllListeners();
-          };
-
-          BufferedSender.prototype.close = function() {
-            debug('close');
-            this._cleanup();
-            if (this.sendStop) {
-              this.sendStop();
-              this.sendStop = null;
-            }
-          };
-
-          module.exports = BufferedSender;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      26: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            IframeTransport = require('../iframe'),
-            objectUtils = require('../../utils/object');
-
-          module.exports = function(transport) {
-
-            function IframeWrapTransport(transUrl, baseUrl) {
-              IframeTransport.call(this, transport.transportName, transUrl, baseUrl);
-            }
-
-            inherits(IframeWrapTransport, IframeTransport);
-
-            IframeWrapTransport.enabled = function(url, info) {
-              if (!global.document) {
-                return false;
-              }
-
-              var iframeInfo = objectUtils.extend({}, info);
-              iframeInfo.sameOrigin = true;
-              return transport.enabled(iframeInfo) && IframeTransport.enabled();
-            };
-
-            IframeWrapTransport.transportName = 'iframe-' + transport.transportName;
-            IframeWrapTransport.needBody = true;
-            IframeWrapTransport.roundTrips = IframeTransport.roundTrips + transport.roundTrips - 1; // html, javascript (2) + transport - no CORS (1)
-
-            IframeWrapTransport.facadeTransport = transport;
-
-            return IframeWrapTransport;
-          };
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../../utils/object": 49,
-        "../iframe": 22,
-        "inherits": 57
-      }],
-      27: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            EventEmitter = require('events').EventEmitter;
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:polling');
-          }
-
-          function Polling(Receiver, receiveUrl, AjaxObject) {
-            debug(receiveUrl);
-            EventEmitter.call(this);
-            this.Receiver = Receiver;
-            this.receiveUrl = receiveUrl;
-            this.AjaxObject = AjaxObject;
-            this._scheduleReceiver();
-          }
-
-          inherits(Polling, EventEmitter);
-
-          Polling.prototype._scheduleReceiver = function() {
-            debug('_scheduleReceiver');
-            var self = this;
-            var poll = this.poll = new this.Receiver(this.receiveUrl, this.AjaxObject);
-
-            poll.on('message', function(msg) {
-              debug('message', msg);
-              self.emit('message', msg);
-            });
-
-            poll.once('close', function(code, reason) {
-              debug('close', code, reason, self.pollIsClosing);
-              self.poll = poll = null;
-
-              if (!self.pollIsClosing) {
-                if (reason === 'network') {
-                  self._scheduleReceiver();
-                } else {
-                  self.emit('close', code || 1006, reason);
-                  self.removeAllListeners();
-                }
-              }
-            });
-          };
-
-          Polling.prototype.abort = function() {
-            debug('abort');
-            this.removeAllListeners();
-            this.pollIsClosing = true;
-            if (this.poll) {
-              this.poll.abort();
-            }
-          };
-
-          module.exports = Polling;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      28: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            urlUtils = require('../../utils/url'),
-            BufferedSender = require('./buffered-sender'),
-            Polling = require('./polling');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:sender-receiver');
-          }
-
-          function SenderReceiver(transUrl, urlSuffix, senderFunc, Receiver, AjaxObject) {
-            var pollUrl = urlUtils.addPath(transUrl, urlSuffix);
-            debug(pollUrl);
-            var self = this;
-            BufferedSender.call(this, transUrl, senderFunc);
-
-            this.poll = new Polling(Receiver, pollUrl, AjaxObject);
-            this.poll.on('message', function(msg) {
-              debug('poll message', msg);
-              self.emit('message', msg);
-            });
-            this.poll.once('close', function(code, reason) {
-              debug('poll close', code, reason);
-              self.poll = null;
-              self.emit('close', code, reason);
-              self.close();
-            });
-          }
-
-          inherits(SenderReceiver, BufferedSender);
-
-          SenderReceiver.prototype.close = function() {
-            BufferedSender.prototype.close.call(this);
-            debug('close');
-            this.removeAllListeners();
-            if (this.poll) {
-              this.poll.abort();
-              this.poll = null;
-            }
-          };
-
-          module.exports = SenderReceiver;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "../../utils/url": 52,
-        "./buffered-sender": 25,
-        "./polling": 27,
-        "debug": 55,
-        "inherits": 57
-      }],
-      29: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            EventEmitter = require('events').EventEmitter,
-            EventSourceDriver = require('eventsource');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:receiver:eventsource');
-          }
-
-          function EventSourceReceiver(url) {
-            debug(url);
-            EventEmitter.call(this);
-
-            var self = this;
-            var es = this.es = new EventSourceDriver(url);
-            es.onmessage = function(e) {
-              debug('message', e.data);
-              self.emit('message', decodeURI(e.data));
-            };
-            es.onerror = function(e) {
-              debug('error', es.readyState, e);
-              // ES on reconnection has readyState = 0 or 1.
-              // on network error it's CLOSED = 2
-              var reason = (es.readyState !== 2 ? 'network' : 'permanent');
-              self._cleanup();
-              self._close(reason);
-            };
-          }
-
-          inherits(EventSourceReceiver, EventEmitter);
-
-          EventSourceReceiver.prototype.abort = function() {
-            debug('abort');
-            this._cleanup();
-            this._close('user');
-          };
-
-          EventSourceReceiver.prototype._cleanup = function() {
-            debug('cleanup');
-            var es = this.es;
-            if (es) {
-              es.onmessage = es.onerror = null;
-              es.close();
-              this.es = null;
-            }
-          };
-
-          EventSourceReceiver.prototype._close = function(reason) {
-            debug('close', reason);
-            var self = this;
-            // Safari and chrome < 15 crash if we close window before
-            // waiting for ES cleanup. See:
-            // https://code.google.com/p/chromium/issues/detail?id=89155
-            setTimeout(function() {
-              self.emit('close', null, reason);
-              self.removeAllListeners();
-            }, 200);
-          };
-
-          module.exports = EventSourceReceiver;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "debug": 55,
-        "events": 3,
-        "eventsource": 18,
-        "inherits": 57
-      }],
-      30: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            iframeUtils = require('../../utils/iframe'),
-            urlUtils = require('../../utils/url'),
-            EventEmitter = require('events').EventEmitter,
-            random = require('../../utils/random');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:receiver:htmlfile');
-          }
-
-          function HtmlfileReceiver(url) {
-            debug(url);
-            EventEmitter.call(this);
-            var self = this;
-            iframeUtils.polluteGlobalNamespace();
-
-            this.id = 'a' + random.string(6);
-            url = urlUtils.addQuery(url, 'c=' + decodeURIComponent(iframeUtils.WPrefix + '.' + this.id));
-
-            debug('using htmlfile', HtmlfileReceiver.htmlfileEnabled);
-            var constructFunc = HtmlfileReceiver.htmlfileEnabled ?
-              iframeUtils.createHtmlfile : iframeUtils.createIframe;
-
-            global[iframeUtils.WPrefix][this.id] = {
-              start: function() {
-                debug('start');
-                self.iframeObj.loaded();
-              },
-              message: function(data) {
-                debug('message', data);
-                self.emit('message', data);
-              },
-              stop: function() {
-                debug('stop');
-                self._cleanup();
-                self._close('network');
-              }
-            };
-            this.iframeObj = constructFunc(url, function() {
-              debug('callback');
-              self._cleanup();
-              self._close('permanent');
-            });
-          }
-
-          inherits(HtmlfileReceiver, EventEmitter);
-
-          HtmlfileReceiver.prototype.abort = function() {
-            debug('abort');
-            this._cleanup();
-            this._close('user');
-          };
-
-          HtmlfileReceiver.prototype._cleanup = function() {
-            debug('_cleanup');
-            if (this.iframeObj) {
-              this.iframeObj.cleanup();
-              this.iframeObj = null;
-            }
-            delete global[iframeUtils.WPrefix][this.id];
-          };
-
-          HtmlfileReceiver.prototype._close = function(reason) {
-            debug('_close', reason);
-            this.emit('close', null, reason);
-            this.removeAllListeners();
-          };
-
-          HtmlfileReceiver.htmlfileEnabled = false;
-
-          // obfuscate to avoid firewalls
-          var axo = ['Active'].concat('Object').join('X');
-          if (axo in global) {
-            try {
-              HtmlfileReceiver.htmlfileEnabled = !!new global[axo]('htmlfile');
-            } catch (x) {
-              // intentionally empty
-            }
-          }
-
-          HtmlfileReceiver.enabled = HtmlfileReceiver.htmlfileEnabled || iframeUtils.iframeEnabled;
-
-          module.exports = HtmlfileReceiver;
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../../utils/iframe": 47,
-        "../../utils/random": 50,
-        "../../utils/url": 52,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      31: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var utils = require('../../utils/iframe'),
-            random = require('../../utils/random'),
-            browser = require('../../utils/browser'),
-            urlUtils = require('../../utils/url'),
-            inherits = require('inherits'),
-            EventEmitter = require('events').EventEmitter;
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:receiver:jsonp');
-          }
-
-          function JsonpReceiver(url) {
-            debug(url);
-            var self = this;
-            EventEmitter.call(this);
-
-            utils.polluteGlobalNamespace();
-
-            this.id = 'a' + random.string(6);
-            var urlWithId = urlUtils.addQuery(url, 'c=' + encodeURIComponent(utils.WPrefix + '.' + this.id));
-
-            global[utils.WPrefix][this.id] = this._callback.bind(this);
-            this._createScript(urlWithId);
-
-            // Fallback mostly for Konqueror - stupid timer, 35 seconds shall be plenty.
-            this.timeoutId = setTimeout(function() {
-              debug('timeout');
-              self._abort(new Error('JSONP script loaded abnormally (timeout)'));
-            }, JsonpReceiver.timeout);
-          }
-
-          inherits(JsonpReceiver, EventEmitter);
-
-          JsonpReceiver.prototype.abort = function() {
-            debug('abort');
-            if (global[utils.WPrefix][this.id]) {
-              var err = new Error('JSONP user aborted read');
-              err.code = 1000;
-              this._abort(err);
-            }
-          };
-
-          JsonpReceiver.timeout = 35000;
-          JsonpReceiver.scriptErrorTimeout = 1000;
-
-          JsonpReceiver.prototype._callback = function(data) {
-            debug('_callback', data);
-            this._cleanup();
-
-            if (this.aborting) {
-              return;
-            }
-
-            if (data) {
-              debug('message', data);
-              this.emit('message', data);
-            }
-            this.emit('close', null, 'network');
-            this.removeAllListeners();
-          };
-
-          JsonpReceiver.prototype._abort = function(err) {
-            debug('_abort', err);
-            this._cleanup();
-            this.aborting = true;
-            this.emit('close', err.code, err.message);
-            this.removeAllListeners();
-          };
-
-          JsonpReceiver.prototype._cleanup = function() {
-            debug('_cleanup');
-            clearTimeout(this.timeoutId);
-            if (this.script2) {
-              this.script2.parentNode.removeChild(this.script2);
-              this.script2 = null;
-            }
-            if (this.script) {
-              var script = this.script;
-              // Unfortunately, you can't really abort script loading of
-              // the script.
-              script.parentNode.removeChild(script);
-              script.onreadystatechange = script.onerror =
-                script.onload = script.onclick = null;
-              this.script = null;
-            }
-            delete global[utils.WPrefix][this.id];
-          };
-
-          JsonpReceiver.prototype._scriptError = function() {
-            debug('_scriptError');
-            var self = this;
-            if (this.errorTimer) {
-              return;
-            }
-
-            this.errorTimer = setTimeout(function() {
-              if (!self.loadedOkay) {
-                self._abort(new Error('JSONP script loaded abnormally (onerror)'));
-              }
-            }, JsonpReceiver.scriptErrorTimeout);
-          };
-
-          JsonpReceiver.prototype._createScript = function(url) {
-            debug('_createScript', url);
-            var self = this;
-            var script = this.script = global.document.createElement('script');
-            var script2; // Opera synchronous load trick.
-
-            script.id = 'a' + random.string(8);
-            script.src = url;
-            script.type = 'text/javascript';
-            script.charset = 'UTF-8';
-            script.onerror = this._scriptError.bind(this);
-            script.onload = function() {
-              debug('onload');
-              self._abort(new Error('JSONP script loaded abnormally (onload)'));
-            };
-
-            // IE9 fires 'error' event after onreadystatechange or before, in random order.
-            // Use loadedOkay to determine if actually errored
-            script.onreadystatechange = function() {
-              debug('onreadystatechange', script.readyState);
-              if (/loaded|closed/.test(script.readyState)) {
-                if (script && script.htmlFor && script.onclick) {
-                  self.loadedOkay = true;
-                  try {
-                    // In IE, actually execute the script.
-                    script.onclick();
-                  } catch (x) {
-                    // intentionally empty
-                  }
-                }
-                if (script) {
-                  self._abort(new Error('JSONP script loaded abnormally (onreadystatechange)'));
-                }
-              }
-            };
-            // IE: event/htmlFor/onclick trick.
-            // One can't rely on proper order for onreadystatechange. In order to
-            // make sure, set a 'htmlFor' and 'event' properties, so that
-            // script code will be installed as 'onclick' handler for the
-            // script object. Later, onreadystatechange, manually execute this
-            // code. FF and Chrome doesn't work with 'event' and 'htmlFor'
-            // set. For reference see:
-            //   http://jaubourg.net/2010/07/loading-script-as-onclick-handler-of.html
-            // Also, read on that about script ordering:
-            //   http://wiki.whatwg.org/wiki/Dynamic_Script_Execution_Order
-            if (typeof script.async === 'undefined' && global.document.attachEvent) {
-              // According to mozilla docs, in recent browsers script.async defaults
-              // to 'true', so we may use it to detect a good browser:
-              // https://developer.mozilla.org/en/HTML/Element/script
-              if (!browser.isOpera()) {
-                // Naively assume we're in IE
-                try {
-                  script.htmlFor = script.id;
-                  script.event = 'onclick';
-                } catch (x) {
-                  // intentionally empty
-                }
-                script.async = true;
-              } else {
-                // Opera, second sync script hack
-                script2 = this.script2 = global.document.createElement('script');
-                script2.text = "try{var a = document.getElementById('" + script.id + "'); if(a)a.onerror();}catch(x){};";
-                script.async = script2.async = false;
-              }
-            }
-            if (typeof script.async !== 'undefined') {
-              script.async = true;
-            }
-
-            var head = global.document.getElementsByTagName('head')[0];
-            head.insertBefore(script, head.firstChild);
-            if (script2) {
-              head.insertBefore(script2, head.firstChild);
-            }
-          };
-
-          module.exports = JsonpReceiver;
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../../utils/browser": 44,
-        "../../utils/iframe": 47,
-        "../../utils/random": 50,
-        "../../utils/url": 52,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      32: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            EventEmitter = require('events').EventEmitter;
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:receiver:xhr');
-          }
-
-          function XhrReceiver(url, AjaxObject) {
-            debug(url);
-            EventEmitter.call(this);
-            var self = this;
-
-            this.bufferPosition = 0;
-
-            this.xo = new AjaxObject('POST', url, null);
-            this.xo.on('chunk', this._chunkHandler.bind(this));
-            this.xo.once('finish', function(status, text) {
-              debug('finish', status, text);
-              self._chunkHandler(status, text);
-              self.xo = null;
-              var reason = status === 200 ? 'network' : 'permanent';
-              debug('close', reason);
-              self.emit('close', null, reason);
-              self._cleanup();
-            });
-          }
-
-          inherits(XhrReceiver, EventEmitter);
-
-          XhrReceiver.prototype._chunkHandler = function(status, text) {
-            debug('_chunkHandler', status);
-            if (status !== 200 || !text) {
-              return;
-            }
-
-            for (var idx = -1;; this.bufferPosition += idx + 1) {
-              var buf = text.slice(this.bufferPosition);
-              idx = buf.indexOf('\n');
-              if (idx === -1) {
-                break;
-              }
-              var msg = buf.slice(0, idx);
-              if (msg) {
-                debug('message', msg);
-                this.emit('message', msg);
-              }
-            }
-          };
-
-          XhrReceiver.prototype._cleanup = function() {
-            debug('_cleanup');
-            this.removeAllListeners();
-          };
-
-          XhrReceiver.prototype.abort = function() {
-            debug('abort');
-            if (this.xo) {
-              this.xo.close();
-              debug('close');
-              this.emit('close', null, 'user');
-              this.xo = null;
-            }
-            this._cleanup();
-          };
-
-          module.exports = XhrReceiver;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      33: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var random = require('../../utils/random'),
-            urlUtils = require('../../utils/url');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:sender:jsonp');
-          }
-
-          var form, area;
-
-          function createIframe(id) {
-            debug('createIframe', id);
-            try {
-              // ie6 dynamic iframes with target="" support (thanks Chris Lambacher)
-              return global.document.createElement('<iframe name="' + id + '">');
-            } catch (x) {
-              var iframe = global.document.createElement('iframe');
-              iframe.name = id;
-              return iframe;
-            }
-          }
-
-          function createForm() {
-            debug('createForm');
-            form = global.document.createElement('form');
-            form.style.display = 'none';
-            form.style.position = 'absolute';
-            form.method = 'POST';
-            form.enctype = 'application/x-www-form-urlencoded';
-            form.acceptCharset = 'UTF-8';
-
-            area = global.document.createElement('textarea');
-            area.name = 'd';
-            form.appendChild(area);
-
-            global.document.body.appendChild(form);
-          }
-
-          module.exports = function(url, payload, callback) {
-            debug(url, payload);
-            if (!form) {
-              createForm();
-            }
-            var id = 'a' + random.string(8);
-            form.target = id;
-            form.action = urlUtils.addQuery(urlUtils.addPath(url, '/jsonp_send'), 'i=' + id);
-
-            var iframe = createIframe(id);
-            iframe.id = id;
-            iframe.style.display = 'none';
-            form.appendChild(iframe);
-
-            try {
-              area.value = payload;
-            } catch (e) {
-              // seriously broken browsers get here
-            }
-            form.submit();
-
-            var completed = function(err) {
-              debug('completed', id, err);
-              if (!iframe.onerror) {
-                return;
-              }
-              iframe.onreadystatechange = iframe.onerror = iframe.onload = null;
-              // Opera mini doesn't like if we GC iframe
-              // immediately, thus this timeout.
-              setTimeout(function() {
-                debug('cleaning up', id);
-                iframe.parentNode.removeChild(iframe);
-                iframe = null;
-              }, 500);
-              area.value = '';
-              // It is not possible to detect if the iframe succeeded or
-              // failed to submit our form.
-              callback(err);
-            };
-            iframe.onerror = function() {
-              debug('onerror', id);
-              completed();
-            };
-            iframe.onload = function() {
-              debug('onload', id);
-              completed();
-            };
-            iframe.onreadystatechange = function(e) {
-              debug('onreadystatechange', id, iframe.readyState, e);
-              if (iframe.readyState === 'complete') {
-                completed();
-              }
-            };
-            return function() {
-              debug('aborted', id);
-              completed(new Error('Aborted'));
-            };
-          };
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../../utils/random": 50,
-        "../../utils/url": 52,
-        "debug": 55
-      }],
-      34: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var EventEmitter = require('events').EventEmitter,
-            inherits = require('inherits'),
-            eventUtils = require('../../utils/event'),
-            browser = require('../../utils/browser'),
-            urlUtils = require('../../utils/url');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:sender:xdr');
-          }
-
-          // References:
-          //   http://ajaxian.com/archives/100-line-ajax-wrapper
-          //   http://msdn.microsoft.com/en-us/library/cc288060(v=VS.85).aspx
-
-          function XDRObject(method, url, payload) {
-            debug(method, url);
-            var self = this;
-            EventEmitter.call(this);
-
-            setTimeout(function() {
-              self._start(method, url, payload);
-            }, 0);
-          }
-
-          inherits(XDRObject, EventEmitter);
-
-          XDRObject.prototype._start = function(method, url, payload) {
-            debug('_start');
-            var self = this;
-            var xdr = new global.XDomainRequest();
-            // IE caches even POSTs
-            url = urlUtils.addQuery(url, 't=' + (+new Date()));
-
-            xdr.onerror = function() {
-              debug('onerror');
-              self._error();
-            };
-            xdr.ontimeout = function() {
-              debug('ontimeout');
-              self._error();
-            };
-            xdr.onprogress = function() {
-              debug('progress', xdr.responseText);
-              self.emit('chunk', 200, xdr.responseText);
-            };
-            xdr.onload = function() {
-              debug('load');
-              self.emit('finish', 200, xdr.responseText);
-              self._cleanup(false);
-            };
-            this.xdr = xdr;
-            this.unloadRef = eventUtils.unloadAdd(function() {
-              self._cleanup(true);
-            });
-            try {
-              // Fails with AccessDenied if port number is bogus
-              this.xdr.open(method, url);
-              if (this.timeout) {
-                this.xdr.timeout = this.timeout;
-              }
-              this.xdr.send(payload);
-            } catch (x) {
-              this._error();
-            }
-          };
-
-          XDRObject.prototype._error = function() {
-            this.emit('finish', 0, '');
-            this._cleanup(false);
-          };
-
-          XDRObject.prototype._cleanup = function(abort) {
-            debug('cleanup', abort);
-            if (!this.xdr) {
-              return;
-            }
-            this.removeAllListeners();
-            eventUtils.unloadDel(this.unloadRef);
-
-            this.xdr.ontimeout = this.xdr.onerror = this.xdr.onprogress = this.xdr.onload = null;
-            if (abort) {
-              try {
-                this.xdr.abort();
-              } catch (x) {
-                // intentionally empty
-              }
-            }
-            this.unloadRef = this.xdr = null;
-          };
-
-          XDRObject.prototype.close = function() {
-            debug('close');
-            this._cleanup(true);
-          };
-
-          // IE 8/9 if the request target uses the same scheme - #79
-          XDRObject.enabled = !!(global.XDomainRequest && browser.hasDomain());
-
-          module.exports = XDRObject;
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../../utils/browser": 44,
-        "../../utils/event": 46,
-        "../../utils/url": 52,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      35: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          XhrDriver = require('../driver/xhr');
-
-        function XHRCorsObject(method, url, payload, opts) {
-          XhrDriver.call(this, method, url, payload, opts);
-        }
-
-        inherits(XHRCorsObject, XhrDriver);
-
-        XHRCorsObject.enabled = XhrDriver.enabled && XhrDriver.supportsCORS;
-
-        module.exports = XHRCorsObject;
-
-      }, {
-        "../driver/xhr": 17,
-        "inherits": 57
-      }],
-      36: [function(require, module, exports) {
-        'use strict';
-
-        var EventEmitter = require('events').EventEmitter,
-          inherits = require('inherits');
-
-        function XHRFake( /* method, url, payload, opts */ ) {
-          var self = this;
-          EventEmitter.call(this);
-
-          this.to = setTimeout(function() {
-            self.emit('finish', 200, '{}');
-          }, XHRFake.timeout);
-        }
-
-        inherits(XHRFake, EventEmitter);
-
-        XHRFake.prototype.close = function() {
-          clearTimeout(this.to);
-        };
-
-        XHRFake.timeout = 2000;
-
-        module.exports = XHRFake;
-
-      }, {
-        "events": 3,
-        "inherits": 57
-      }],
-      37: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          XhrDriver = require('../driver/xhr');
-
-        function XHRLocalObject(method, url, payload /*, opts */ ) {
-          XhrDriver.call(this, method, url, payload, {
-            noCredentials: true
-          });
-        }
-
-        inherits(XHRLocalObject, XhrDriver);
-
-        XHRLocalObject.enabled = XhrDriver.enabled;
-
-        module.exports = XHRLocalObject;
-
-      }, {
-        "../driver/xhr": 17,
-        "inherits": 57
-      }],
-      38: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var utils = require('../utils/event'),
-            urlUtils = require('../utils/url'),
-            inherits = require('inherits'),
-            EventEmitter = require('events').EventEmitter,
-            WebsocketDriver = require('./driver/websocket');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:websocket');
-          }
-
-          function WebSocketTransport(transUrl, ignore, options) {
-            if (!WebSocketTransport.enabled()) {
-              throw new Error('Transport created when disabled');
-            }
-
-            EventEmitter.call(this);
-            debug('constructor', transUrl);
-
-            var self = this;
-            var url = urlUtils.addPath(transUrl, '/websocket');
-            if (url.slice(0, 5) === 'https') {
-              url = 'wss' + url.slice(5);
-            } else {
-              url = 'ws' + url.slice(4);
-            }
-            this.url = url;
-
-            this.ws = new WebsocketDriver(this.url, [], options);
-            this.ws.onmessage = function(e) {
-              debug('message event', e.data);
-              self.emit('message', e.data);
-            };
-            // Firefox has an interesting bug. If a websocket connection is
-            // created after onunload, it stays alive even when user
-            // navigates away from the page. In such situation let's lie -
-            // let's not open the ws connection at all. See:
-            // https://github.com/sockjs/sockjs-client/issues/28
-            // https://bugzilla.mozilla.org/show_bug.cgi?id=696085
-            this.unloadRef = utils.unloadAdd(function() {
-              debug('unload');
-              self.ws.close();
-            });
-            this.ws.onclose = function(e) {
-              debug('close event', e.code, e.reason);
-              self.emit('close', e.code, e.reason);
-              self._cleanup();
-            };
-            this.ws.onerror = function(e) {
-              debug('error event', e);
-              self.emit('close', 1006, 'WebSocket connection broken');
-              self._cleanup();
-            };
-          }
-
-          inherits(WebSocketTransport, EventEmitter);
-
-          WebSocketTransport.prototype.send = function(data) {
-            var msg = '[' + data + ']';
-            debug('send', msg);
-            this.ws.send(msg);
-          };
-
-          WebSocketTransport.prototype.close = function() {
-            debug('close');
-            var ws = this.ws;
-            this._cleanup();
-            if (ws) {
-              ws.close();
-            }
-          };
-
-          WebSocketTransport.prototype._cleanup = function() {
-            debug('_cleanup');
-            var ws = this.ws;
-            if (ws) {
-              ws.onmessage = ws.onclose = ws.onerror = null;
-            }
-            utils.unloadDel(this.unloadRef);
-            this.unloadRef = this.ws = null;
-            this.removeAllListeners();
-          };
-
-          WebSocketTransport.enabled = function() {
-            debug('enabled');
-            return !!WebsocketDriver;
-          };
-          WebSocketTransport.transportName = 'websocket';
-
-          // In theory, ws should require 1 round trip. But in chrome, this is
-          // not very stable over SSL. Most likely a ws connection requires a
-          // separate SSL connection, in which case 2 round trips are an
-          // absolute minumum.
-          WebSocketTransport.roundTrips = 2;
-
-          module.exports = WebSocketTransport;
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "../utils/event": 46,
-        "../utils/url": 52,
-        "./driver/websocket": 19,
-        "debug": 55,
-        "events": 3,
-        "inherits": 57
-      }],
-      39: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          AjaxBasedTransport = require('./lib/ajax-based'),
-          XdrStreamingTransport = require('./xdr-streaming'),
-          XhrReceiver = require('./receiver/xhr'),
-          XDRObject = require('./sender/xdr');
-
-        function XdrPollingTransport(transUrl) {
-          if (!XDRObject.enabled) {
-            throw new Error('Transport created when disabled');
-          }
-          AjaxBasedTransport.call(this, transUrl, '/xhr', XhrReceiver, XDRObject);
-        }
-
-        inherits(XdrPollingTransport, AjaxBasedTransport);
-
-        XdrPollingTransport.enabled = XdrStreamingTransport.enabled;
-        XdrPollingTransport.transportName = 'xdr-polling';
-        XdrPollingTransport.roundTrips = 2; // preflight, ajax
-
-        module.exports = XdrPollingTransport;
-
-      }, {
-        "./lib/ajax-based": 24,
-        "./receiver/xhr": 32,
-        "./sender/xdr": 34,
-        "./xdr-streaming": 40,
-        "inherits": 57
-      }],
-      40: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          AjaxBasedTransport = require('./lib/ajax-based'),
-          XhrReceiver = require('./receiver/xhr'),
-          XDRObject = require('./sender/xdr');
-
-        // According to:
-        //   http://stackoverflow.com/questions/1641507/detect-browser-support-for-cross-domain-xmlhttprequests
-        //   http://hacks.mozilla.org/2009/07/cross-site-xmlhttprequest-with-cors/
-
-        function XdrStreamingTransport(transUrl) {
-          if (!XDRObject.enabled) {
-            throw new Error('Transport created when disabled');
-          }
-          AjaxBasedTransport.call(this, transUrl, '/xhr_streaming', XhrReceiver, XDRObject);
-        }
-
-        inherits(XdrStreamingTransport, AjaxBasedTransport);
-
-        XdrStreamingTransport.enabled = function(info) {
-          if (info.cookie_needed || info.nullOrigin) {
-            return false;
-          }
-          return XDRObject.enabled && info.sameScheme;
-        };
-
-        XdrStreamingTransport.transportName = 'xdr-streaming';
-        XdrStreamingTransport.roundTrips = 2; // preflight, ajax
-
-        module.exports = XdrStreamingTransport;
-
-      }, {
-        "./lib/ajax-based": 24,
-        "./receiver/xhr": 32,
-        "./sender/xdr": 34,
-        "inherits": 57
-      }],
-      41: [function(require, module, exports) {
-        'use strict';
-
-        var inherits = require('inherits'),
-          AjaxBasedTransport = require('./lib/ajax-based'),
-          XhrReceiver = require('./receiver/xhr'),
-          XHRCorsObject = require('./sender/xhr-cors'),
-          XHRLocalObject = require('./sender/xhr-local');
-
-        function XhrPollingTransport(transUrl) {
-          if (!XHRLocalObject.enabled && !XHRCorsObject.enabled) {
-            throw new Error('Transport created when disabled');
-          }
-          AjaxBasedTransport.call(this, transUrl, '/xhr', XhrReceiver, XHRCorsObject);
-        }
-
-        inherits(XhrPollingTransport, AjaxBasedTransport);
-
-        XhrPollingTransport.enabled = function(info) {
-          if (info.nullOrigin) {
-            return false;
-          }
-
-          if (XHRLocalObject.enabled && info.sameOrigin) {
-            return true;
-          }
-          return XHRCorsObject.enabled;
-        };
-
-        XhrPollingTransport.transportName = 'xhr-polling';
-        XhrPollingTransport.roundTrips = 2; // preflight, ajax
-
-        module.exports = XhrPollingTransport;
-
-      }, {
-        "./lib/ajax-based": 24,
-        "./receiver/xhr": 32,
-        "./sender/xhr-cors": 35,
-        "./sender/xhr-local": 37,
-        "inherits": 57
-      }],
-      42: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var inherits = require('inherits'),
-            AjaxBasedTransport = require('./lib/ajax-based'),
-            XhrReceiver = require('./receiver/xhr'),
-            XHRCorsObject = require('./sender/xhr-cors'),
-            XHRLocalObject = require('./sender/xhr-local'),
-            browser = require('../utils/browser');
-
-          function XhrStreamingTransport(transUrl) {
-            if (!XHRLocalObject.enabled && !XHRCorsObject.enabled) {
-              throw new Error('Transport created when disabled');
-            }
-            AjaxBasedTransport.call(this, transUrl, '/xhr_streaming', XhrReceiver, XHRCorsObject);
-          }
-
-          inherits(XhrStreamingTransport, AjaxBasedTransport);
-
-          XhrStreamingTransport.enabled = function(info) {
-            if (info.nullOrigin) {
-              return false;
-            }
-            // Opera doesn't support xhr-streaming #60
-            // But it might be able to #92
-            if (browser.isOpera()) {
-              return false;
-            }
-
-            return XHRCorsObject.enabled;
-          };
-
-          XhrStreamingTransport.transportName = 'xhr-streaming';
-          XhrStreamingTransport.roundTrips = 2; // preflight, ajax
-
-          // Safari gets confused when a streaming ajax request is started
-          // before onload. This causes the load indicator to spin indefinetely.
-          // Only require body when used in a browser
-          XhrStreamingTransport.needBody = !!global.document;
-
-          module.exports = XhrStreamingTransport;
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "../utils/browser": 44,
-        "./lib/ajax-based": 24,
-        "./receiver/xhr": 32,
-        "./sender/xhr-cors": 35,
-        "./sender/xhr-local": 37,
-        "inherits": 57
-      }],
-      43: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          if (global.crypto && global.crypto.getRandomValues) {
-            module.exports.randomBytes = function(length) {
-              var bytes = new Uint8Array(length);
-              global.crypto.getRandomValues(bytes);
-              return bytes;
-            };
-          } else {
-            module.exports.randomBytes = function(length) {
-              var bytes = new Array(length);
-              for (var i = 0; i < length; i++) {
-                bytes[i] = Math.floor(Math.random() * 256);
-              }
-              return bytes;
-            };
-          }
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      44: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          module.exports = {
-            isOpera: function() {
-              return global.navigator &&
-                /opera/i.test(global.navigator.userAgent);
-            }
-
-            ,
-            isKonqueror: function() {
-              return global.navigator &&
-                /konqueror/i.test(global.navigator.userAgent);
-            }
-
-            // #187 wrap document.domain in try/catch because of WP8 from file:///
-            ,
-            hasDomain: function() {
-              // non-browser client always has a domain
-              if (!global.document) {
-                return true;
-              }
-
-              try {
-                return !!global.document.domain;
-              } catch (e) {
-                return false;
-              }
-            }
-          };
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      45: [function(require, module, exports) {
-        'use strict';
-
-        var JSON3 = require('json3');
-
-        // Some extra characters that Chrome gets wrong, and substitutes with
-        // something else on the wire.
-        // eslint-disable-next-line no-control-regex
-        var extraEscapable = /[\x00-\x1f\ud800-\udfff\ufffe\uffff\u0300-\u0333\u033d-\u0346\u034a-\u034c\u0350-\u0352\u0357-\u0358\u035c-\u0362\u0374\u037e\u0387\u0591-\u05af\u05c4\u0610-\u0617\u0653-\u0654\u0657-\u065b\u065d-\u065e\u06df-\u06e2\u06eb-\u06ec\u0730\u0732-\u0733\u0735-\u0736\u073a\u073d\u073f-\u0741\u0743\u0745\u0747\u07eb-\u07f1\u0951\u0958-\u095f\u09dc-\u09dd\u09df\u0a33\u0a36\u0a59-\u0a5b\u0a5e\u0b5c-\u0b5d\u0e38-\u0e39\u0f43\u0f4d\u0f52\u0f57\u0f5c\u0f69\u0f72-\u0f76\u0f78\u0f80-\u0f83\u0f93\u0f9d\u0fa2\u0fa7\u0fac\u0fb9\u1939-\u193a\u1a17\u1b6b\u1cda-\u1cdb\u1dc0-\u1dcf\u1dfc\u1dfe\u1f71\u1f73\u1f75\u1f77\u1f79\u1f7b\u1f7d\u1fbb\u1fbe\u1fc9\u1fcb\u1fd3\u1fdb\u1fe3\u1feb\u1fee-\u1fef\u1ff9\u1ffb\u1ffd\u2000-\u2001\u20d0-\u20d1\u20d4-\u20d7\u20e7-\u20e9\u2126\u212a-\u212b\u2329-\u232a\u2adc\u302b-\u302c\uaab2-\uaab3\uf900-\ufa0d\ufa10\ufa12\ufa15-\ufa1e\ufa20\ufa22\ufa25-\ufa26\ufa2a-\ufa2d\ufa30-\ufa6d\ufa70-\ufad9\ufb1d\ufb1f\ufb2a-\ufb36\ufb38-\ufb3c\ufb3e\ufb40-\ufb41\ufb43-\ufb44\ufb46-\ufb4e\ufff0-\uffff]/g,
-          extraLookup;
-
-        // This may be quite slow, so let's delay until user actually uses bad
-        // characters.
-        var unrollLookup = function(escapable) {
-          var i;
-          var unrolled = {};
-          var c = [];
-          for (i = 0; i < 65536; i++) {
-            c.push(String.fromCharCode(i));
-          }
-          escapable.lastIndex = 0;
-          c.join('').replace(escapable, function(a) {
-            unrolled[a] = '\\u' + ('0000' + a.charCodeAt(0).toString(16)).slice(-4);
-            return '';
-          });
-          escapable.lastIndex = 0;
-          return unrolled;
-        };
-
-        // Quote string, also taking care of unicode characters that browsers
-        // often break. Especially, take care of unicode surrogates:
-        // http://en.wikipedia.org/wiki/Mapping_of_Unicode_characters#Surrogates
-        module.exports = {
-          quote: function(string) {
-            var quoted = JSON3.stringify(string);
-
-            // In most cases this should be very fast and good enough.
-            extraEscapable.lastIndex = 0;
-            if (!extraEscapable.test(quoted)) {
-              return quoted;
-            }
-
-            if (!extraLookup) {
-              extraLookup = unrollLookup(extraEscapable);
-            }
-
-            return quoted.replace(extraEscapable, function(a) {
-              return extraLookup[a];
-            });
-          }
-        };
-
-      }, {
-        "json3": 58
-      }],
-      46: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var random = require('./random');
-
-          var onUnload = {},
-            afterUnload = false
-            // detect google chrome packaged apps because they don't allow the 'unload' event
-            ,
-            isChromePackagedApp = global.chrome && global.chrome.app && global.chrome.app.runtime;
-
-          module.exports = {
-            attachEvent: function(event, listener) {
-              if (typeof global.addEventListener !== 'undefined') {
-                global.addEventListener(event, listener, false);
-              } else if (global.document && global.attachEvent) {
-                // IE quirks.
-                // According to: http://stevesouders.com/misc/test-postmessage.php
-                // the message gets delivered only to 'document', not 'window'.
-                global.document.attachEvent('on' + event, listener);
-                // I get 'window' for ie8.
-                global.attachEvent('on' + event, listener);
-              }
-            }
-
-            ,
-            detachEvent: function(event, listener) {
-              if (typeof global.addEventListener !== 'undefined') {
-                global.removeEventListener(event, listener, false);
-              } else if (global.document && global.detachEvent) {
-                global.document.detachEvent('on' + event, listener);
-                global.detachEvent('on' + event, listener);
-              }
-            }
-
-            ,
-            unloadAdd: function(listener) {
-              if (isChromePackagedApp) {
-                return null;
-              }
-
-              var ref = random.string(8);
-              onUnload[ref] = listener;
-              if (afterUnload) {
-                setTimeout(this.triggerUnloadCallbacks, 0);
-              }
-              return ref;
-            }
-
-            ,
-            unloadDel: function(ref) {
-              if (ref in onUnload) {
-                delete onUnload[ref];
-              }
-            }
-
-            ,
-            triggerUnloadCallbacks: function() {
-              for (var ref in onUnload) {
-                onUnload[ref]();
-                delete onUnload[ref];
-              }
-            }
-          };
-
-          var unloadTriggered = function() {
-            if (afterUnload) {
-              return;
-            }
-            afterUnload = true;
-            module.exports.triggerUnloadCallbacks();
-          };
-
-          // 'unload' alone is not reliable in opera within an iframe, but we
-          // can't use `beforeunload` as IE fires it on javascript: links.
-          if (!isChromePackagedApp) {
-            module.exports.attachEvent('unload', unloadTriggered);
-          }
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./random": 50
-      }],
-      47: [function(require, module, exports) {
-        (function(process, global) {
-          'use strict';
-
-          var eventUtils = require('./event'),
-            JSON3 = require('json3'),
-            browser = require('./browser');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:utils:iframe');
-          }
-
-          module.exports = {
-            WPrefix: '_jp',
-            currentWindowId: null
-
-            ,
-            polluteGlobalNamespace: function() {
-              if (!(module.exports.WPrefix in global)) {
-                global[module.exports.WPrefix] = {};
-              }
-            }
-
-            ,
-            postMessage: function(type, data) {
-              if (global.parent !== global) {
-                global.parent.postMessage(JSON3.stringify({
-                  windowId: module.exports.currentWindowId,
-                  type: type,
-                  data: data || ''
-                }), '*');
-              } else {
-                debug('Cannot postMessage, no parent window.', type, data);
-              }
-            }
-
-            ,
-            createIframe: function(iframeUrl, errorCallback) {
-              var iframe = global.document.createElement('iframe');
-              var tref, unloadRef;
-              var unattach = function() {
-                debug('unattach');
-                clearTimeout(tref);
-                // Explorer had problems with that.
-                try {
-                  iframe.onload = null;
-                } catch (x) {
-                  // intentionally empty
-                }
-                iframe.onerror = null;
-              };
-              var cleanup = function() {
-                debug('cleanup');
-                if (iframe) {
-                  unattach();
-                  // This timeout makes chrome fire onbeforeunload event
-                  // within iframe. Without the timeout it goes straight to
-                  // onunload.
-                  setTimeout(function() {
-                    if (iframe) {
-                      iframe.parentNode.removeChild(iframe);
-                    }
-                    iframe = null;
-                  }, 0);
-                  eventUtils.unloadDel(unloadRef);
-                }
-              };
-              var onerror = function(err) {
-                debug('onerror', err);
-                if (iframe) {
-                  cleanup();
-                  errorCallback(err);
-                }
-              };
-              var post = function(msg, origin) {
-                debug('post', msg, origin);
-                try {
-                  // When the iframe is not loaded, IE raises an exception
-                  // on 'contentWindow'.
-                  setTimeout(function() {
-                    if (iframe && iframe.contentWindow) {
-                      iframe.contentWindow.postMessage(msg, origin);
-                    }
-                  }, 0);
-                } catch (x) {
-                  // intentionally empty
-                }
-              };
-
-              iframe.src = iframeUrl;
-              iframe.style.display = 'none';
-              iframe.style.position = 'absolute';
-              iframe.onerror = function() {
-                onerror('onerror');
-              };
-              iframe.onload = function() {
-                debug('onload');
-                // `onload` is triggered before scripts on the iframe are
-                // executed. Give it few seconds to actually load stuff.
-                clearTimeout(tref);
-                tref = setTimeout(function() {
-                  onerror('onload timeout');
-                }, 2000);
-              };
-              global.document.body.appendChild(iframe);
-              tref = setTimeout(function() {
-                onerror('timeout');
-              }, 15000);
-              unloadRef = eventUtils.unloadAdd(cleanup);
-              return {
-                post: post,
-                cleanup: cleanup,
-                loaded: unattach
-              };
-            }
-
-            /* eslint no-undef: "off", new-cap: "off" */
-            ,
-            createHtmlfile: function(iframeUrl, errorCallback) {
-              var axo = ['Active'].concat('Object').join('X');
-              var doc = new global[axo]('htmlfile');
-              var tref, unloadRef;
-              var iframe;
-              var unattach = function() {
-                clearTimeout(tref);
-                iframe.onerror = null;
-              };
-              var cleanup = function() {
-                if (doc) {
-                  unattach();
-                  eventUtils.unloadDel(unloadRef);
-                  iframe.parentNode.removeChild(iframe);
-                  iframe = doc = null;
-                  CollectGarbage();
-                }
-              };
-              var onerror = function(r) {
-                debug('onerror', r);
-                if (doc) {
-                  cleanup();
-                  errorCallback(r);
-                }
-              };
-              var post = function(msg, origin) {
-                try {
-                  // When the iframe is not loaded, IE raises an exception
-                  // on 'contentWindow'.
-                  setTimeout(function() {
-                    if (iframe && iframe.contentWindow) {
-                      iframe.contentWindow.postMessage(msg, origin);
-                    }
-                  }, 0);
-                } catch (x) {
-                  // intentionally empty
-                }
-              };
-
-              doc.open();
-              doc.write('<html><s' + 'cript>' +
-                'document.domain="' + global.document.domain + '";' +
-                '</s' + 'cript></html>');
-              doc.close();
-              doc.parentWindow[module.exports.WPrefix] = global[module.exports.WPrefix];
-              var c = doc.createElement('div');
-              doc.body.appendChild(c);
-              iframe = doc.createElement('iframe');
-              c.appendChild(iframe);
-              iframe.src = iframeUrl;
-              iframe.onerror = function() {
-                onerror('onerror');
-              };
-              tref = setTimeout(function() {
-                onerror('timeout');
-              }, 15000);
-              unloadRef = eventUtils.unloadAdd(cleanup);
-              return {
-                post: post,
-                cleanup: cleanup,
-                loaded: unattach
-              };
-            }
-          };
-
-          module.exports.iframeEnabled = false;
-          if (global.document) {
-            // postMessage misbehaves in konqueror 4.6.5 - the messages are delivered with
-            // huge delay, or not at all.
-            module.exports.iframeEnabled = (typeof global.postMessage === 'function' ||
-              typeof global.postMessage === 'object') && (!browser.isKonqueror());
-          }
-
-        }).call(this, {
-          env: {}
-        }, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./browser": 44,
-        "./event": 46,
-        "debug": 55,
-        "json3": 58
-      }],
-      48: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var logObject = {};
-          ['log', 'debug', 'warn'].forEach(function(level) {
-            var levelExists;
-
-            try {
-              levelExists = global.console && global.console[level] && global.console[level].apply;
-            } catch (e) {
-              // do nothing
-            }
-
-            logObject[level] = levelExists ? function() {
-              return global.console[level].apply(global.console, arguments);
-            } : (level === 'log' ? function() {} : logObject.log);
-          });
-
-          module.exports = logObject;
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      49: [function(require, module, exports) {
-        'use strict';
-
-        module.exports = {
-          isObject: function(obj) {
-            var type = typeof obj;
-            return type === 'function' || type === 'object' && !!obj;
-          }
-
-          ,
-          extend: function(obj) {
-            if (!this.isObject(obj)) {
-              return obj;
-            }
-            var source, prop;
-            for (var i = 1, length = arguments.length; i < length; i++) {
-              source = arguments[i];
-              for (prop in source) {
-                if (Object.prototype.hasOwnProperty.call(source, prop)) {
-                  obj[prop] = source[prop];
-                }
-              }
-            }
-            return obj;
-          }
-        };
-
-      }, {}],
-      50: [function(require, module, exports) {
-        'use strict';
-
-        /* global crypto:true */
-        var crypto = require('crypto');
-
-        // This string has length 32, a power of 2, so the modulus doesn't introduce a
-        // bias.
-        var _randomStringChars = 'abcdefghijklmnopqrstuvwxyz012345';
-        module.exports = {
-          string: function(length) {
-            var max = _randomStringChars.length;
-            var bytes = crypto.randomBytes(length);
-            var ret = [];
-            for (var i = 0; i < length; i++) {
-              ret.push(_randomStringChars.substr(bytes[i] % max, 1));
-            }
-            return ret.join('');
-          }
-
-          ,
-          number: function(max) {
-            return Math.floor(Math.random() * max);
-          }
-
-          ,
-          numberString: function(max) {
-            var t = ('' + (max - 1)).length;
-            var p = new Array(t + 1).join('0');
-            return (p + this.number(max)).slice(-t);
-          }
-        };
-
-      }, {
-        "crypto": 43
-      }],
-      51: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:utils:transport');
-          }
-
-          module.exports = function(availableTransports) {
-            return {
-              filterToEnabled: function(transportsWhitelist, info) {
-                var transports = {
-                  main: [],
-                  facade: []
-                };
-                if (!transportsWhitelist) {
-                  transportsWhitelist = [];
-                } else if (typeof transportsWhitelist === 'string') {
-                  transportsWhitelist = [transportsWhitelist];
-                }
-
-                availableTransports.forEach(function(trans) {
-                  if (!trans) {
-                    return;
-                  }
-
-                  if (trans.transportName === 'websocket' && info.websocket === false) {
-                    debug('disabled from server', 'websocket');
-                    return;
-                  }
-
-                  if (transportsWhitelist.length &&
-                    transportsWhitelist.indexOf(trans.transportName) === -1) {
-                    debug('not in whitelist', trans.transportName);
-                    return;
-                  }
-
-                  if (trans.enabled(info)) {
-                    debug('enabled', trans.transportName);
-                    transports.main.push(trans);
-                    if (trans.facadeTransport) {
-                      transports.facade.push(trans.facadeTransport);
-                    }
-                  } else {
-                    debug('disabled', trans.transportName);
-                  }
-                });
-                return transports;
-              }
-            };
-          };
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "debug": 55
-      }],
-      52: [function(require, module, exports) {
-        (function(process) {
-          'use strict';
-
-          var URL = require('url-parse');
-
-          var debug = function() {};
-          if (process.env.NODE_ENV !== 'production') {
-            debug = require('debug')('sockjs-client:utils:url');
-          }
-
-          module.exports = {
-            getOrigin: function(url) {
-              if (!url) {
-                return null;
-              }
-
-              var p = new URL(url);
-              if (p.protocol === 'file:') {
-                return null;
-              }
-
-              var port = p.port;
-              if (!port) {
-                port = (p.protocol === 'https:') ? '443' : '80';
-              }
-
-              return p.protocol + '//' + p.hostname + ':' + port;
-            }
-
-            ,
-            isOriginEqual: function(a, b) {
-              var res = this.getOrigin(a) === this.getOrigin(b);
-              debug('same', a, b, res);
-              return res;
-            }
-
-            ,
-            isSchemeEqual: function(a, b) {
-              return (a.split(':')[0] === b.split(':')[0]);
-            }
-
-            ,
-            addPath: function(url, path) {
-              var qs = url.split('?');
-              return qs[0] + path + (qs[1] ? '?' + qs[1] : '');
-            }
-
-            ,
-            addQuery: function(url, q) {
-              return url + (url.indexOf('?') === -1 ? ('?' + q) : ('&' + q));
-            }
-          };
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "debug": 55,
-        "url-parse": 61
-      }],
-      53: [function(require, module, exports) {
-        module.exports = '1.1.4';
-
-      }, {}],
-      54: [function(require, module, exports) {
-        /**
-         * Helpers.
-         */
-
-        var s = 1000
-        var m = s * 60
-        var h = m * 60
-        var d = h * 24
-        var y = d * 365.25
-
-        /**
-         * Parse or format the given `val`.
-         *
-         * Options:
-         *
-         *  - `long` verbose formatting [false]
-         *
-         * @param {String|Number} val
-         * @param {Object} [options]
-         * @throws {Error} throw an error if val is not a non-empty string or a number
-         * @return {String|Number}
-         * @api public
-         */
-
-        module.exports = function(val, options) {
-          options = options || {}
-          var type = typeof val
-          if (type === 'string' && val.length > 0) {
-            return parse(val)
-          } else if (type === 'number' && isNaN(val) === false) {
-            return options.long ?
-              fmtLong(val) :
-              fmtShort(val)
-          }
-          throw new Error('val is not a non-empty string or a valid number. val=' + JSON.stringify(val))
-        }
-
-        /**
-         * Parse the given `str` and return milliseconds.
-         *
-         * @param {String} str
-         * @return {Number}
-         * @api private
-         */
-
-        function parse(str) {
-          str = String(str)
-          if (str.length > 10000) {
-            return
-          }
-          var match = /^((?:\d+)?\.?\d+) *(milliseconds?|msecs?|ms|seconds?|secs?|s|minutes?|mins?|m|hours?|hrs?|h|days?|d|years?|yrs?|y)?$/i.exec(str)
-          if (!match) {
-            return
-          }
-          var n = parseFloat(match[1])
-          var type = (match[2] || 'ms').toLowerCase()
-          switch (type) {
-            case 'years':
-            case 'year':
-            case 'yrs':
-            case 'yr':
-            case 'y':
-              return n * y
-            case 'days':
-            case 'day':
-            case 'd':
-              return n * d
-            case 'hours':
-            case 'hour':
-            case 'hrs':
-            case 'hr':
-            case 'h':
-              return n * h
-            case 'minutes':
-            case 'minute':
-            case 'mins':
-            case 'min':
-            case 'm':
-              return n * m
-            case 'seconds':
-            case 'second':
-            case 'secs':
-            case 'sec':
-            case 's':
-              return n * s
-            case 'milliseconds':
-            case 'millisecond':
-            case 'msecs':
-            case 'msec':
-            case 'ms':
-              return n
-            default:
-              return undefined
-          }
-        }
-
-        /**
-         * Short format for `ms`.
-         *
-         * @param {Number} ms
-         * @return {String}
-         * @api private
-         */
-
-        function fmtShort(ms) {
-          if (ms >= d) {
-            return Math.round(ms / d) + 'd'
-          }
-          if (ms >= h) {
-            return Math.round(ms / h) + 'h'
-          }
-          if (ms >= m) {
-            return Math.round(ms / m) + 'm'
-          }
-          if (ms >= s) {
-            return Math.round(ms / s) + 's'
-          }
-          return ms + 'ms'
-        }
-
-        /**
-         * Long format for `ms`.
-         *
-         * @param {Number} ms
-         * @return {String}
-         * @api private
-         */
-
-        function fmtLong(ms) {
-          return plural(ms, d, 'day') ||
-            plural(ms, h, 'hour') ||
-            plural(ms, m, 'minute') ||
-            plural(ms, s, 'second') ||
-            ms + ' ms'
-        }
-
-        /**
-         * Pluralization helper.
-         */
-
-        function plural(ms, n, name) {
-          if (ms < n) {
-            return
-          }
-          if (ms < n * 1.5) {
-            return Math.floor(ms / n) + ' ' + name
-          }
-          return Math.ceil(ms / n) + ' ' + name + 's'
-        }
-
-      }, {}],
-      55: [function(require, module, exports) {
-        (function(process) {
-          /**
-           * This is the web browser implementation of `debug()`.
-           *
-           * Expose `debug()` as the module.
-           */
-
-          exports = module.exports = require('./debug');
-          exports.log = log;
-          exports.formatArgs = formatArgs;
-          exports.save = save;
-          exports.load = load;
-          exports.useColors = useColors;
-          exports.storage = 'undefined' != typeof chrome &&
-            'undefined' != typeof chrome.storage ?
-            chrome.storage.local :
-            localstorage();
-
-          /**
-           * Colors.
-           */
-
-          exports.colors = [
-            'lightseagreen',
-            'forestgreen',
-            'goldenrod',
-            'dodgerblue',
-            'darkorchid',
-            'crimson'
-          ];
-
-          /**
-           * Currently only WebKit-based Web Inspectors, Firefox >= v31,
-           * and the Firebug extension (any Firefox version) are known
-           * to support "%c" CSS customizations.
-           *
-           * TODO: add a `localStorage` variable to explicitly enable/disable colors
-           */
-
-          function useColors() {
-            // NB: In an Electron preload script, document will be defined but not fully
-            // initialized. Since we know we're in Chrome, we'll just detect this case
-            // explicitly
-            if (typeof window !== 'undefined' && window.process && window.process.type === 'renderer') {
-              return true;
-            }
-
-            // is webkit? http://stackoverflow.com/a/16459606/376773
-            // document is undefined in react-native: https://github.com/facebook/react-native/pull/1632
-            return (typeof document !== 'undefined' && document && document.documentElement && document.documentElement.style && document.documentElement.style.WebkitAppearance) ||
-              // is firebug? http://stackoverflow.com/a/398120/376773
-              (typeof window !== 'undefined' && window && window.console && (window.console.firebug || (window.console.exception && window.console.table))) ||
-              // is firefox >= v31?
-              // https://developer.mozilla.org/en-US/docs/Tools/Web_Console#Styling_messages
-              (typeof navigator !== 'undefined' && navigator && navigator.userAgent && navigator.userAgent.toLowerCase().match(/firefox\/(\d+)/) && parseInt(RegExp.$1, 10) >= 31) ||
-              // double check webkit in userAgent just in case we are in a worker
-              (typeof navigator !== 'undefined' && navigator && navigator.userAgent && navigator.userAgent.toLowerCase().match(/applewebkit\/(\d+)/));
-          }
-
-          /**
-           * Map %j to `JSON.stringify()`, since no Web Inspectors do that by default.
-           */
-
-          exports.formatters.j = function(v) {
-            try {
-              return JSON.stringify(v);
-            } catch (err) {
-              return '[UnexpectedJSONParseError]: ' + err.message;
-            }
-          };
-
-
-          /**
-           * Colorize log arguments if enabled.
-           *
-           * @api public
-           */
-
-          function formatArgs(args) {
-            var useColors = this.useColors;
-
-            args[0] = (useColors ? '%c' : '') +
-              this.namespace +
-              (useColors ? ' %c' : ' ') +
-              args[0] +
-              (useColors ? '%c ' : ' ') +
-              '+' + exports.humanize(this.diff);
-
-            if (!useColors) return;
-
-            var c = 'color: ' + this.color;
-            args.splice(1, 0, c, 'color: inherit')
-
-            // the final "%c" is somewhat tricky, because there could be other
-            // arguments passed either before or after the %c, so we need to
-            // figure out the correct index to insert the CSS into
-            var index = 0;
-            var lastC = 0;
-            args[0].replace(/%[a-zA-Z%]/g, function(match) {
-              if ('%%' === match) return;
-              index++;
-              if ('%c' === match) {
-                // we only are interested in the *last* %c
-                // (the user may have provided their own)
-                lastC = index;
-              }
-            });
-
-            args.splice(lastC, 0, c);
-          }
-
-          /**
-           * Invokes `console.log()` when available.
-           * No-op when `console.log` is not a "function".
-           *
-           * @api public
-           */
-
-          function log() {
-            // this hackery is required for IE8/9, where
-            // the `console.log` function doesn't have 'apply'
-            return 'object' === typeof console &&
-              console.log &&
-              Function.prototype.apply.call(console.log, console, arguments);
-          }
-
-          /**
-           * Save `namespaces`.
-           *
-           * @param {String} namespaces
-           * @api private
-           */
-
-          function save(namespaces) {
-            try {
-              if (null == namespaces) {
-                exports.storage.removeItem('debug');
-              } else {
-                exports.storage.debug = namespaces;
-              }
-            } catch (e) {}
-          }
-
-          /**
-           * Load `namespaces`.
-           *
-           * @return {String} returns the previously persisted debug modes
-           * @api private
-           */
-
-          function load() {
-            var r;
-            try {
-              r = exports.storage.debug;
-            } catch (e) {}
-
-            // If debug isn't set in LS, and we're in Electron, try to load $DEBUG
-            if (!r && typeof process !== 'undefined' && 'env' in process) {
-              r = process.env.DEBUG;
-            }
-
-            return r;
-          }
-
-          /**
-           * Enable namespaces listed in `localStorage.debug` initially.
-           */
-
-          exports.enable(load());
-
-          /**
-           * Localstorage attempts to return the localstorage.
-           *
-           * This is necessary because safari throws
-           * when a user disables cookies/localstorage
-           * and you attempt to access it.
-           *
-           * @return {LocalStorage}
-           * @api private
-           */
-
-          function localstorage() {
-            try {
-              return window.localStorage;
-            } catch (e) {}
-          }
-
-        }).call(this, {
-          env: {}
-        })
-
-      }, {
-        "./debug": 56
-      }],
-      56: [function(require, module, exports) {
-
-        /**
-         * This is the common logic for both the Node.js and web browser
-         * implementations of `debug()`.
-         *
-         * Expose `debug()` as the module.
-         */
-
-        exports = module.exports = createDebug.debug = createDebug['default'] = createDebug;
-        exports.coerce = coerce;
-        exports.disable = disable;
-        exports.enable = enable;
-        exports.enabled = enabled;
-        exports.humanize = require('ms');
-
-        /**
-         * The currently active debug mode names, and names to skip.
-         */
-
-        exports.names = [];
-        exports.skips = [];
-
-        /**
-         * Map of special "%n" handling functions, for the debug "format" argument.
-         *
-         * Valid key names are a single, lower or upper-case letter, i.e. "n" and "N".
-         */
-
-        exports.formatters = {};
-
-        /**
-         * Previous log timestamp.
-         */
-
-        var prevTime;
-
-        /**
-         * Select a color.
-         * @param {String} namespace
-         * @return {Number}
-         * @api private
-         */
-
-        function selectColor(namespace) {
-          var hash = 0,
-            i;
-
-          for (i in namespace) {
-            hash = ((hash << 5) - hash) + namespace.charCodeAt(i);
-            hash |= 0; // Convert to 32bit integer
-          }
-
-          return exports.colors[Math.abs(hash) % exports.colors.length];
-        }
-
-        /**
-         * Create a debugger with the given `namespace`.
-         *
-         * @param {String} namespace
-         * @return {Function}
-         * @api public
-         */
-
-        function createDebug(namespace) {
-
-          function debug() {
-            // disabled?
-            if (!debug.enabled) return;
-
-            var self = debug;
-
-            // set `diff` timestamp
-            var curr = +new Date();
-            var ms = curr - (prevTime || curr);
-            self.diff = ms;
-            self.prev = prevTime;
-            self.curr = curr;
-            prevTime = curr;
-
-            // turn the `arguments` into a proper Array
-            var args = new Array(arguments.length);
-            for (var i = 0; i < args.length; i++) {
-              args[i] = arguments[i];
-            }
-
-            args[0] = exports.coerce(args[0]);
-
-            if ('string' !== typeof args[0]) {
-              // anything else let's inspect with %O
-              args.unshift('%O');
-            }
-
-            // apply any `formatters` transformations
-            var index = 0;
-            args[0] = args[0].replace(/%([a-zA-Z%])/g, function(match, format) {
-              // if we encounter an escaped % then don't increase the array index
-              if (match === '%%') return match;
-              index++;
-              var formatter = exports.formatters[format];
-              if ('function' === typeof formatter) {
-                var val = args[index];
-                match = formatter.call(self, val);
-
-                // now we need to remove `args[index]` since it's inlined in the `format`
-                args.splice(index, 1);
-                index--;
-              }
-              return match;
-            });
-
-            // apply env-specific formatting (colors, etc.)
-            exports.formatArgs.call(self, args);
-
-            var logFn = debug.log || exports.log || console.log.bind(console);
-            logFn.apply(self, args);
-          }
-
-          debug.namespace = namespace;
-          debug.enabled = exports.enabled(namespace);
-          debug.useColors = exports.useColors();
-          debug.color = selectColor(namespace);
-
-          // env-specific initialization logic for debug instances
-          if ('function' === typeof exports.init) {
-            exports.init(debug);
-          }
-
-          return debug;
-        }
-
-        /**
-         * Enables a debug mode by namespaces. This can include modes
-         * separated by a colon and wildcards.
-         *
-         * @param {String} namespaces
-         * @api public
-         */
-
-        function enable(namespaces) {
-          exports.save(namespaces);
-
-          exports.names = [];
-          exports.skips = [];
-
-          var split = (typeof namespaces === 'string' ? namespaces : '').split(/[\s,]+/);
-          var len = split.length;
-
-          for (var i = 0; i < len; i++) {
-            if (!split[i]) continue; // ignore empty strings
-            namespaces = split[i].replace(/\*/g, '.*?');
-            if (namespaces[0] === '-') {
-              exports.skips.push(new RegExp('^' + namespaces.substr(1) + '$'));
-            } else {
-              exports.names.push(new RegExp('^' + namespaces + '$'));
-            }
-          }
-        }
-
-        /**
-         * Disable debug output.
-         *
-         * @api public
-         */
-
-        function disable() {
-          exports.enable('');
-        }
-
-        /**
-         * Returns true if the given mode name is enabled, false otherwise.
-         *
-         * @param {String} name
-         * @return {Boolean}
-         * @api public
-         */
-
-        function enabled(name) {
-          var i, len;
-          for (i = 0, len = exports.skips.length; i < len; i++) {
-            if (exports.skips[i].test(name)) {
-              return false;
-            }
-          }
-          for (i = 0, len = exports.names.length; i < len; i++) {
-            if (exports.names[i].test(name)) {
-              return true;
-            }
-          }
-          return false;
-        }
-
-        /**
-         * Coerce `val`.
-         *
-         * @param {Mixed} val
-         * @return {Mixed}
-         * @api private
-         */
-
-        function coerce(val) {
-          if (val instanceof Error) return val.stack || val.message;
-          return val;
-        }
-
-      }, {
-        "ms": 54
-      }],
-      57: [function(require, module, exports) {
-        if (typeof Object.create === 'function') {
-          // implementation from standard node.js 'util' module
-          module.exports = function inherits(ctor, superCtor) {
-            ctor.super_ = superCtor
-            ctor.prototype = Object.create(superCtor.prototype, {
-              constructor: {
-                value: ctor,
-                enumerable: false,
-                writable: true,
-                configurable: true
-              }
-            });
-          };
-        } else {
-          // old school shim for old browsers
-          module.exports = function inherits(ctor, superCtor) {
-            ctor.super_ = superCtor
-            var TempCtor = function() {}
-            TempCtor.prototype = superCtor.prototype
-            ctor.prototype = new TempCtor()
-            ctor.prototype.constructor = ctor
-          }
-        }
-
-      }, {}],
-      58: [function(require, module, exports) {
-        (function(global) {
-          /*! JSON v3.3.2 | http://bestiejs.github.io/json3 | Copyright 2012-2014, Kit Cambridge | http://kit.mit-license.org */
-          ;
-          (function() {
-            // Detect the `define` function exposed by asynchronous module loaders. The
-            // strict `define` check is necessary for compatibility with `r.js`.
-            var isLoader = typeof define === "function" && define.amd;
-
-            // A set of types used to distinguish objects from primitives.
-            var objectTypes = {
-              "function": true,
-              "object": true
-            };
-
-            // Detect the `exports` object exposed by CommonJS implementations.
-            var freeExports = objectTypes[typeof exports] && exports && !exports.nodeType && exports;
-
-            // Use the `global` object exposed by Node (including Browserify via
-            // `insert-module-globals`), Narwhal, and Ringo as the default context,
-            // and the `window` object in browsers. Rhino exports a `global` function
-            // instead.
-            var root = objectTypes[typeof window] && window || this,
-              freeGlobal = freeExports && objectTypes[typeof module] && module && !module.nodeType && typeof global == "object" && global;
-
-            if (freeGlobal && (freeGlobal["global"] === freeGlobal || freeGlobal["window"] === freeGlobal || freeGlobal["self"] === freeGlobal)) {
-              root = freeGlobal;
-            }
-
-            // Public: Initializes JSON 3 using the given `context` object, attaching the
-            // `stringify` and `parse` functions to the specified `exports` object.
-            function runInContext(context, exports) {
-              context || (context = root["Object"]());
-              exports || (exports = root["Object"]());
-
-              // Native constructor aliases.
-              var Number = context["Number"] || root["Number"],
-                String = context["String"] || root["String"],
-                Object = context["Object"] || root["Object"],
-                Date = context["Date"] || root["Date"],
-                SyntaxError = context["SyntaxError"] || root["SyntaxError"],
-                TypeError = context["TypeError"] || root["TypeError"],
-                Math = context["Math"] || root["Math"],
-                nativeJSON = context["JSON"] || root["JSON"];
-
-              // Delegate to the native `stringify` and `parse` implementations.
-              if (typeof nativeJSON == "object" && nativeJSON) {
-                exports.stringify = nativeJSON.stringify;
-                exports.parse = nativeJSON.parse;
-              }
-
-              // Convenience aliases.
-              var objectProto = Object.prototype,
-                getClass = objectProto.toString,
-                isProperty, forEach, undef;
-
-              // Test the `Date#getUTC*` methods. Based on work by @Yaffle.
-              var isExtended = new Date(-3509827334573292);
-              try {
-                // The `getUTCFullYear`, `Month`, and `Date` methods return nonsensical
-                // results for certain dates in Opera >= 10.53.
-                isExtended = isExtended.getUTCFullYear() == -109252 && isExtended.getUTCMonth() === 0 && isExtended.getUTCDate() === 1 &&
-                  // Safari < 2.0.2 stores the internal millisecond time value correctly,
-                  // but clips the values returned by the date methods to the range of
-                  // signed 32-bit integers ([-2 ** 31, 2 ** 31 - 1]).
-                  isExtended.getUTCHours() == 10 && isExtended.getUTCMinutes() == 37 && isExtended.getUTCSeconds() == 6 && isExtended.getUTCMilliseconds() == 708;
-              } catch (exception) {}
-
-              // Internal: Determines whether the native `JSON.stringify` and `parse`
-              // implementations are spec-compliant. Based on work by Ken Snyder.
-              function has(name) {
-                if (has[name] !== undef) {
-                  // Return cached feature test result.
-                  return has[name];
-                }
-                var isSupported;
-                if (name == "bug-string-char-index") {
-                  // IE <= 7 doesn't support accessing string characters using square
-                  // bracket notation. IE 8 only supports this for primitives.
-                  isSupported = "a" [0] != "a";
-                } else if (name == "json") {
-                  // Indicates whether both `JSON.stringify` and `JSON.parse` are
-                  // supported.
-                  isSupported = has("json-stringify") && has("json-parse");
-                } else {
-                  var value, serialized = '{"a":[1,true,false,null,"\\u0000\\b\\n\\f\\r\\t"]}';
-                  // Test `JSON.stringify`.
-                  if (name == "json-stringify") {
-                    var stringify = exports.stringify,
-                      stringifySupported = typeof stringify == "function" && isExtended;
-                    if (stringifySupported) {
-                      // A test function object with a custom `toJSON` method.
-                      (value = function() {
-                        return 1;
-                      }).toJSON = value;
-                      try {
-                        stringifySupported =
-                          // Firefox 3.1b1 and b2 serialize string, number, and boolean
-                          // primitives as object literals.
-                          stringify(0) === "0" &&
-                          // FF 3.1b1, b2, and JSON 2 serialize wrapped primitives as object
-                          // literals.
-                          stringify(new Number()) === "0" &&
-                          stringify(new String()) == '""' &&
-                          // FF 3.1b1, 2 throw an error if the value is `null`, `undefined`, or
-                          // does not define a canonical JSON representation (this applies to
-                          // objects with `toJSON` properties as well, *unless* they are nested
-                          // within an object or array).
-                          stringify(getClass) === undef &&
-                          // IE 8 serializes `undefined` as `"undefined"`. Safari <= 5.1.7 and
-                          // FF 3.1b3 pass this test.
-                          stringify(undef) === undef &&
-                          // Safari <= 5.1.7 and FF 3.1b3 throw `Error`s and `TypeError`s,
-                          // respectively, if the value is omitted entirely.
-                          stringify() === undef &&
-                          // FF 3.1b1, 2 throw an error if the given value is not a number,
-                          // string, array, object, Boolean, or `null` literal. This applies to
-                          // objects with custom `toJSON` methods as well, unless they are nested
-                          // inside object or array literals. YUI 3.0.0b1 ignores custom `toJSON`
-                          // methods entirely.
-                          stringify(value) === "1" &&
-                          stringify([value]) == "[1]" &&
-                          // Prototype <= 1.6.1 serializes `[undefined]` as `"[]"` instead of
-                          // `"[null]"`.
-                          stringify([undef]) == "[null]" &&
-                          // YUI 3.0.0b1 fails to serialize `null` literals.
-                          stringify(null) == "null" &&
-                          // FF 3.1b1, 2 halts serialization if an array contains a function:
-                          // `[1, true, getClass, 1]` serializes as "[1,true,],". FF 3.1b3
-                          // elides non-JSON values from objects and arrays, unless they
-                          // define custom `toJSON` methods.
-                          stringify([undef, getClass, null]) == "[null,null,null]" &&
-                          // Simple serialization test. FF 3.1b1 uses Unicode escape sequences
-                          // where character escape codes are expected (e.g., `\b` => `\u0008`).
-                          stringify({
-                            "a": [value, true, false, null, "\x00\b\n\f\r\t"]
-                          }) == serialized &&
-                          // FF 3.1b1 and b2 ignore the `filter` and `width` arguments.
-                          stringify(null, value) === "1" &&
-                          stringify([1, 2], null, 1) == "[\n 1,\n 2\n]" &&
-                          // JSON 2, Prototype <= 1.7, and older WebKit builds incorrectly
-                          // serialize extended years.
-                          stringify(new Date(-8.64e15)) == '"-271821-04-20T00:00:00.000Z"' &&
-                          // The milliseconds are optional in ES 5, but required in 5.1.
-                          stringify(new Date(8.64e15)) == '"+275760-09-13T00:00:00.000Z"' &&
-                          // Firefox <= 11.0 incorrectly serializes years prior to 0 as negative
-                          // four-digit years instead of six-digit years. Credits: @Yaffle.
-                          stringify(new Date(-621987552e5)) == '"-000001-01-01T00:00:00.000Z"' &&
-                          // Safari <= 5.1.5 and Opera >= 10.53 incorrectly serialize millisecond
-                          // values less than 1000. Credits: @Yaffle.
-                          stringify(new Date(-1)) == '"1969-12-31T23:59:59.999Z"';
-                      } catch (exception) {
-                        stringifySupported = false;
-                      }
-                    }
-                    isSupported = stringifySupported;
-                  }
-                  // Test `JSON.parse`.
-                  if (name == "json-parse") {
-                    var parse = exports.parse;
-                    if (typeof parse == "function") {
-                      try {
-                        // FF 3.1b1, b2 will throw an exception if a bare literal is provided.
-                        // Conforming implementations should also coerce the initial argument to
-                        // a string prior to parsing.
-                        if (parse("0") === 0 && !parse(false)) {
-                          // Simple parsing test.
-                          value = parse(serialized);
-                          var parseSupported = value["a"].length == 5 && value["a"][0] === 1;
-                          if (parseSupported) {
-                            try {
-                              // Safari <= 5.1.2 and FF 3.1b1 allow unescaped tabs in strings.
-                              parseSupported = !parse('"\t"');
-                            } catch (exception) {}
-                            if (parseSupported) {
-                              try {
-                                // FF 4.0 and 4.0.1 allow leading `+` signs and leading
-                                // decimal points. FF 4.0, 4.0.1, and IE 9-10 also allow
-                                // certain octal literals.
-                                parseSupported = parse("01") !== 1;
-                              } catch (exception) {}
-                            }
-                            if (parseSupported) {
-                              try {
-                                // FF 4.0, 4.0.1, and Rhino 1.7R3-R4 allow trailing decimal
-                                // points. These environments, along with FF 3.1b1 and 2,
-                                // also allow trailing commas in JSON objects and arrays.
-                                parseSupported = parse("1.") !== 1;
-                              } catch (exception) {}
-                            }
-                          }
-                        }
-                      } catch (exception) {
-                        parseSupported = false;
-                      }
-                    }
-                    isSupported = parseSupported;
-                  }
-                }
-                return has[name] = !!isSupported;
-              }
-
-              if (!has("json")) {
-                // Common `[[Class]]` name aliases.
-                var functionClass = "[object Function]",
-                  dateClass = "[object Date]",
-                  numberClass = "[object Number]",
-                  stringClass = "[object String]",
-                  arrayClass = "[object Array]",
-                  booleanClass = "[object Boolean]";
-
-                // Detect incomplete support for accessing string characters by index.
-                var charIndexBuggy = has("bug-string-char-index");
-
-                // Define additional utility methods if the `Date` methods are buggy.
-                if (!isExtended) {
-                  var floor = Math.floor;
-                  // A mapping between the months of the year and the number of days between
-                  // January 1st and the first of the respective month.
-                  var Months = [0, 31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334];
-                  // Internal: Calculates the number of days between the Unix epoch and the
-                  // first day of the given month.
-                  var getDay = function(year, month) {
-                    return Months[month] + 365 * (year - 1970) + floor((year - 1969 + (month = +(month > 1))) / 4) - floor((year - 1901 + month) / 100) + floor((year - 1601 + month) / 400);
-                  };
-                }
-
-                // Internal: Determines if a property is a direct property of the given
-                // object. Delegates to the native `Object#hasOwnProperty` method.
-                if (!(isProperty = objectProto.hasOwnProperty)) {
-                  isProperty = function(property) {
-                    var members = {},
-                      constructor;
-                    if ((members.__proto__ = null, members.__proto__ = {
-                        // The *proto* property cannot be set multiple times in recent
-                        // versions of Firefox and SeaMonkey.
-                        "toString": 1
-                      }, members).toString != getClass) {
-                      // Safari <= 2.0.3 doesn't implement `Object#hasOwnProperty`, but
-                      // supports the mutable *proto* property.
-                      isProperty = function(property) {
-                        // Capture and break the object's prototype chain (see section 8.6.2
-                        // of the ES 5.1 spec). The parenthesized expression prevents an
-                        // unsafe transformation by the Closure Compiler.
-                        var original = this.__proto__,
-                          result = property in (this.__proto__ = null, this);
-                        // Restore the original prototype chain.
-                        this.__proto__ = original;
-                        return result;
-                      };
-                    } else {
-                      // Capture a reference to the top-level `Object` constructor.
-                      constructor = members.constructor;
-                      // Use the `constructor` property to simulate `Object#hasOwnProperty` in
-                      // other environments.
-                      isProperty = function(property) {
-                        var parent = (this.constructor || constructor).prototype;
-                        return property in this && !(property in parent && this[property] === parent[property]);
-                      };
-                    }
-                    members = null;
-                    return isProperty.call(this, property);
-                  };
-                }
-
-                // Internal: Normalizes the `for...in` iteration algorithm across
-                // environments. Each enumerated key is yielded to a `callback` function.
-                forEach = function(object, callback) {
-                  var size = 0,
-                    Properties, members, property;
-
-                  // Tests for bugs in the current environment's `for...in` algorithm. The
-                  // `valueOf` property inherits the non-enumerable flag from
-                  // `Object.prototype` in older versions of IE, Netscape, and Mozilla.
-                  (Properties = function() {
-                    this.valueOf = 0;
-                  }).prototype.valueOf = 0;
-
-                  // Iterate over a new instance of the `Properties` class.
-                  members = new Properties();
-                  for (property in members) {
-                    // Ignore all properties inherited from `Object.prototype`.
-                    if (isProperty.call(members, property)) {
-                      size++;
-                    }
-                  }
-                  Properties = members = null;
-
-                  // Normalize the iteration algorithm.
-                  if (!size) {
-                    // A list of non-enumerable properties inherited from `Object.prototype`.
-                    members = ["valueOf", "toString", "toLocaleString", "propertyIsEnumerable", "isPrototypeOf", "hasOwnProperty", "constructor"];
-                    // IE <= 8, Mozilla 1.0, and Netscape 6.2 ignore shadowed non-enumerable
-                    // properties.
-                    forEach = function(object, callback) {
-                      var isFunction = getClass.call(object) == functionClass,
-                        property, length;
-                      var hasProperty = !isFunction && typeof object.constructor != "function" && objectTypes[typeof object.hasOwnProperty] && object.hasOwnProperty || isProperty;
-                      for (property in object) {
-                        // Gecko <= 1.0 enumerates the `prototype` property of functions under
-                        // certain conditions; IE does not.
-                        if (!(isFunction && property == "prototype") && hasProperty.call(object, property)) {
-                          callback(property);
-                        }
-                      }
-                      // Manually invoke the callback for each non-enumerable property.
-                      for (length = members.length; property = members[--length]; hasProperty.call(object, property) && callback(property));
-                    };
-                  } else if (size == 2) {
-                    // Safari <= 2.0.4 enumerates shadowed properties twice.
-                    forEach = function(object, callback) {
-                      // Create a set of iterated properties.
-                      var members = {},
-                        isFunction = getClass.call(object) == functionClass,
-                        property;
-                      for (property in object) {
-                        // Store each property name to prevent double enumeration. The
-                        // `prototype` property of functions is not enumerated due to cross-
-                        // environment inconsistencies.
-                        if (!(isFunction && property == "prototype") && !isProperty.call(members, property) && (members[property] = 1) && isProperty.call(object, property)) {
-                          callback(property);
-                        }
-                      }
-                    };
-                  } else {
-                    // No bugs detected; use the standard `for...in` algorithm.
-                    forEach = function(object, callback) {
-                      var isFunction = getClass.call(object) == functionClass,
-                        property, isConstructor;
-                      for (property in object) {
-                        if (!(isFunction && property == "prototype") && isProperty.call(object, property) && !(isConstructor = property === "constructor")) {
-                          callback(property);
-                        }
-                      }
-                      // Manually invoke the callback for the `constructor` property due to
-                      // cross-environment inconsistencies.
-                      if (isConstructor || isProperty.call(object, (property = "constructor"))) {
-                        callback(property);
-                      }
-                    };
-                  }
-                  return forEach(object, callback);
-                };
-
-                // Public: Serializes a JavaScript `value` as a JSON string. The optional
-                // `filter` argument may specify either a function that alters how object and
-                // array members are serialized, or an array of strings and numbers that
-                // indicates which properties should be serialized. The optional `width`
-                // argument may be either a string or number that specifies the indentation
-                // level of the output.
-                if (!has("json-stringify")) {
-                  // Internal: A map of control characters and their escaped equivalents.
-                  var Escapes = {
-                    92: "\\\\",
-                    34: '\\"',
-                    8: "\\b",
-                    12: "\\f",
-                    10: "\\n",
-                    13: "\\r",
-                    9: "\\t"
-                  };
-
-                  // Internal: Converts `value` into a zero-padded string such that its
-                  // length is at least equal to `width`. The `width` must be <= 6.
-                  var leadingZeroes = "000000";
-                  var toPaddedString = function(width, value) {
-                    // The `|| 0` expression is necessary to work around a bug in
-                    // Opera <= 7.54u2 where `0 == -0`, but `String(-0) !== "0"`.
-                    return (leadingZeroes + (value || 0)).slice(-width);
-                  };
-
-                  // Internal: Double-quotes a string `value`, replacing all ASCII control
-                  // characters (characters with code unit values between 0 and 31) with
-                  // their escaped equivalents. This is an implementation of the
-                  // `Quote(value)` operation defined in ES 5.1 section 15.12.3.
-                  var unicodePrefix = "\\u00";
-                  var quote = function(value) {
-                    var result = '"',
-                      index = 0,
-                      length = value.length,
-                      useCharIndex = !charIndexBuggy || length > 10;
-                    var symbols = useCharIndex && (charIndexBuggy ? value.split("") : value);
-                    for (; index < length; index++) {
-                      var charCode = value.charCodeAt(index);
-                      // If the character is a control character, append its Unicode or
-                      // shorthand escape sequence; otherwise, append the character as-is.
-                      switch (charCode) {
-                        case 8:
-                        case 9:
-                        case 10:
-                        case 12:
-                        case 13:
-                        case 34:
-                        case 92:
-                          result += Escapes[charCode];
-                          break;
-                        default:
-                          if (charCode < 32) {
-                            result += unicodePrefix + toPaddedString(2, charCode.toString(16));
-                            break;
-                          }
-                          result += useCharIndex ? symbols[index] : value.charAt(index);
-                      }
-                    }
-                    return result + '"';
-                  };
-
-                  // Internal: Recursively serializes an object. Implements the
-                  // `Str(key, holder)`, `JO(value)`, and `JA(value)` operations.
-                  var serialize = function(property, object, callback, properties, whitespace, indentation, stack) {
-                    var value, className, year, month, date, time, hours, minutes, seconds, milliseconds, results, element, index, length, prefix, result;
-                    try {
-                      // Necessary for host object support.
-                      value = object[property];
-                    } catch (exception) {}
-                    if (typeof value == "object" && value) {
-                      className = getClass.call(value);
-                      if (className == dateClass && !isProperty.call(value, "toJSON")) {
-                        if (value > -1 / 0 && value < 1 / 0) {
-                          // Dates are serialized according to the `Date#toJSON` method
-                          // specified in ES 5.1 section 15.9.5.44. See section 15.9.1.15
-                          // for the ISO 8601 date time string format.
-                          if (getDay) {
-                            // Manually compute the year, month, date, hours, minutes,
-                            // seconds, and milliseconds if the `getUTC*` methods are
-                            // buggy. Adapted from @Yaffle's `date-shim` project.
-                            date = floor(value / 864e5);
-                            for (year = floor(date / 365.2425) + 1970 - 1; getDay(year + 1, 0) <= date; year++);
-                            for (month = floor((date - getDay(year, 0)) / 30.42); getDay(year, month + 1) <= date; month++);
-                            date = 1 + date - getDay(year, month);
-                            // The `time` value specifies the time within the day (see ES
-                            // 5.1 section 15.9.1.2). The formula `(A % B + B) % B` is used
-                            // to compute `A modulo B`, as the `%` operator does not
-                            // correspond to the `modulo` operation for negative numbers.
-                            time = (value % 864e5 + 864e5) % 864e5;
-                            // The hours, minutes, seconds, and milliseconds are obtained by
-                            // decomposing the time within the day. See section 15.9.1.10.
-                            hours = floor(time / 36e5) % 24;
-                            minutes = floor(time / 6e4) % 60;
-                            seconds = floor(time / 1e3) % 60;
-                            milliseconds = time % 1e3;
-                          } else {
-                            year = value.getUTCFullYear();
-                            month = value.getUTCMonth();
-                            date = value.getUTCDate();
-                            hours = value.getUTCHours();
-                            minutes = value.getUTCMinutes();
-                            seconds = value.getUTCSeconds();
-                            milliseconds = value.getUTCMilliseconds();
-                          }
-                          // Serialize extended years correctly.
-                          value = (year <= 0 || year >= 1e4 ? (year < 0 ? "-" : "+") + toPaddedString(6, year < 0 ? -year : year) : toPaddedString(4, year)) +
-                            "-" + toPaddedString(2, month + 1) + "-" + toPaddedString(2, date) +
-                            // Months, dates, hours, minutes, and seconds should have two
-                            // digits; milliseconds should have three.
-                            "T" + toPaddedString(2, hours) + ":" + toPaddedString(2, minutes) + ":" + toPaddedString(2, seconds) +
-                            // Milliseconds are optional in ES 5.0, but required in 5.1.
-                            "." + toPaddedString(3, milliseconds) + "Z";
-                        } else {
-                          value = null;
-                        }
-                      } else if (typeof value.toJSON == "function" && ((className != numberClass && className != stringClass && className != arrayClass) || isProperty.call(value, "toJSON"))) {
-                        // Prototype <= 1.6.1 adds non-standard `toJSON` methods to the
-                        // `Number`, `String`, `Date`, and `Array` prototypes. JSON 3
-                        // ignores all `toJSON` methods on these objects unless they are
-                        // defined directly on an instance.
-                        value = value.toJSON(property);
-                      }
-                    }
-                    if (callback) {
-                      // If a replacement function was provided, call it to obtain the value
-                      // for serialization.
-                      value = callback.call(object, property, value);
-                    }
-                    if (value === null) {
-                      return "null";
-                    }
-                    className = getClass.call(value);
-                    if (className == booleanClass) {
-                      // Booleans are represented literally.
-                      return "" + value;
-                    } else if (className == numberClass) {
-                      // JSON numbers must be finite. `Infinity` and `NaN` are serialized as
-                      // `"null"`.
-                      return value > -1 / 0 && value < 1 / 0 ? "" + value : "null";
-                    } else if (className == stringClass) {
-                      // Strings are double-quoted and escaped.
-                      return quote("" + value);
-                    }
-                    // Recursively serialize objects and arrays.
-                    if (typeof value == "object") {
-                      // Check for cyclic structures. This is a linear search; performance
-                      // is inversely proportional to the number of unique nested objects.
-                      for (length = stack.length; length--;) {
-                        if (stack[length] === value) {
-                          // Cyclic structures cannot be serialized by `JSON.stringify`.
-                          throw TypeError();
-                        }
-                      }
-                      // Add the object to the stack of traversed objects.
-                      stack.push(value);
-                      results = [];
-                      // Save the current indentation level and indent one additional level.
-                      prefix = indentation;
-                      indentation += whitespace;
-                      if (className == arrayClass) {
-                        // Recursively serialize array elements.
-                        for (index = 0, length = value.length; index < length; index++) {
-                          element = serialize(index, value, callback, properties, whitespace, indentation, stack);
-                          results.push(element === undef ? "null" : element);
-                        }
-                        result = results.length ? (whitespace ? "[\n" + indentation + results.join(",\n" + indentation) + "\n" + prefix + "]" : ("[" + results.join(",") + "]")) : "[]";
-                      } else {
-                        // Recursively serialize object members. Members are selected from
-                        // either a user-specified list of property names, or the object
-                        // itself.
-                        forEach(properties || value, function(property) {
-                          var element = serialize(property, value, callback, properties, whitespace, indentation, stack);
-                          if (element !== undef) {
-                            // According to ES 5.1 section 15.12.3: "If `gap` {whitespace}
-                            // is not the empty string, let `member` {quote(property) + ":"}
-                            // be the concatenation of `member` and the `space` character."
-                            // The "`space` character" refers to the literal space
-                            // character, not the `space` {width} argument provided to
-                            // `JSON.stringify`.
-                            results.push(quote(property) + ":" + (whitespace ? " " : "") + element);
-                          }
-                        });
-                        result = results.length ? (whitespace ? "{\n" + indentation + results.join(",\n" + indentation) + "\n" + prefix + "}" : ("{" + results.join(",") + "}")) : "{}";
-                      }
-                      // Remove the object from the traversed object stack.
-                      stack.pop();
-                      return result;
-                    }
-                  };
-
-                  // Public: `JSON.stringify`. See ES 5.1 section 15.12.3.
-                  exports.stringify = function(source, filter, width) {
-                    var whitespace, callback, properties, className;
-                    if (objectTypes[typeof filter] && filter) {
-                      if ((className = getClass.call(filter)) == functionClass) {
-                        callback = filter;
-                      } else if (className == arrayClass) {
-                        // Convert the property names array into a makeshift set.
-                        properties = {};
-                        for (var index = 0, length = filter.length, value; index < length; value = filter[index++], ((className = getClass.call(value)), className == stringClass || className == numberClass) && (properties[value] = 1));
-                      }
-                    }
-                    if (width) {
-                      if ((className = getClass.call(width)) == numberClass) {
-                        // Convert the `width` to an integer and create a string containing
-                        // `width` number of space characters.
-                        if ((width -= width % 1) > 0) {
-                          for (whitespace = "", width > 10 && (width = 10); whitespace.length < width; whitespace += " ");
-                        }
-                      } else if (className == stringClass) {
-                        whitespace = width.length <= 10 ? width : width.slice(0, 10);
-                      }
-                    }
-                    // Opera <= 7.54u2 discards the values associated with empty string keys
-                    // (`""`) only if they are used directly within an object member list
-                    // (e.g., `!("" in { "": 1})`).
-                    return serialize("", (value = {}, value[""] = source, value), callback, properties, whitespace, "", []);
-                  };
-                }
-
-                // Public: Parses a JSON source string.
-                if (!has("json-parse")) {
-                  var fromCharCode = String.fromCharCode;
-
-                  // Internal: A map of escaped control characters and their unescaped
-                  // equivalents.
-                  var Unescapes = {
-                    92: "\\",
-                    34: '"',
-                    47: "/",
-                    98: "\b",
-                    116: "\t",
-                    110: "\n",
-                    102: "\f",
-                    114: "\r"
-                  };
-
-                  // Internal: Stores the parser state.
-                  var Index, Source;
-
-                  // Internal: Resets the parser state and throws a `SyntaxError`.
-                  var abort = function() {
-                    Index = Source = null;
-                    throw SyntaxError();
-                  };
-
-                  // Internal: Returns the next token, or `"$"` if the parser has reached
-                  // the end of the source string. A token may be a string, number, `null`
-                  // literal, or Boolean literal.
-                  var lex = function() {
-                    var source = Source,
-                      length = source.length,
-                      value, begin, position, isSigned, charCode;
-                    while (Index < length) {
-                      charCode = source.charCodeAt(Index);
-                      switch (charCode) {
-                        case 9:
-                        case 10:
-                        case 13:
-                        case 32:
-                          // Skip whitespace tokens, including tabs, carriage returns, line
-                          // feeds, and space characters.
-                          Index++;
-                          break;
-                        case 123:
-                        case 125:
-                        case 91:
-                        case 93:
-                        case 58:
-                        case 44:
-                          // Parse a punctuator token (`{`, `}`, `[`, `]`, `:`, or `,`) at
-                          // the current position.
-                          value = charIndexBuggy ? source.charAt(Index) : source[Index];
-                          Index++;
-                          return value;
-                        case 34:
-                          // `"` delimits a JSON string; advance to the next character and
-                          // begin parsing the string. String tokens are prefixed with the
-                          // sentinel `@` character to distinguish them from punctuators and
-                          // end-of-string tokens.
-                          for (value = "@", Index++; Index < length;) {
-                            charCode = source.charCodeAt(Index);
-                            if (charCode < 32) {
-                              // Unescaped ASCII control characters (those with a code unit
-                              // less than the space character) are not permitted.
-                              abort();
-                            } else if (charCode == 92) {
-                              // A reverse solidus (`\`) marks the beginning of an escaped
-                              // control character (including `"`, `\`, and `/`) or Unicode
-                              // escape sequence.
-                              charCode = source.charCodeAt(++Index);
-                              switch (charCode) {
-                                case 92:
-                                case 34:
-                                case 47:
-                                case 98:
-                                case 116:
-                                case 110:
-                                case 102:
-                                case 114:
-                                  // Revive escaped control characters.
-                                  value += Unescapes[charCode];
-                                  Index++;
-                                  break;
-                                case 117:
-                                  // `\u` marks the beginning of a Unicode escape sequence.
-                                  // Advance to the first character and validate the
-                                  // four-digit code point.
-                                  begin = ++Index;
-                                  for (position = Index + 4; Index < position; Index++) {
-                                    charCode = source.charCodeAt(Index);
-                                    // A valid sequence comprises four hexdigits (case-
-                                    // insensitive) that form a single hexadecimal value.
-                                    if (!(charCode >= 48 && charCode <= 57 || charCode >= 97 && charCode <= 102 || charCode >= 65 && charCode <= 70)) {
-                                      // Invalid Unicode escape sequence.
-                                      abort();
-                                    }
-                                  }
-                                  // Revive the escaped character.
-                                  value += fromCharCode("0x" + source.slice(begin, Index));
-                                  break;
-                                default:
-                                  // Invalid escape sequence.
-                                  abort();
-                              }
-                            } else {
-                              if (charCode == 34) {
-                                // An unescaped double-quote character marks the end of the
-                                // string.
-                                break;
-                              }
-                              charCode = source.charCodeAt(Index);
-                              begin = Index;
-                              // Optimize for the common case where a string is valid.
-                              while (charCode >= 32 && charCode != 92 && charCode != 34) {
-                                charCode = source.charCodeAt(++Index);
-                              }
-                              // Append the string as-is.
-                              value += source.slice(begin, Index);
-                            }
-                          }
-                          if (source.charCodeAt(Index) == 34) {
-                            // Advance to the next character and return the revived string.
-                            Index++;
-                            return value;
-                          }
-                          // Unterminated string.
-                          abort();
-                        default:
-                          // Parse numbers and literals.
-                          begin = Index;
-                          // Advance past the negative sign, if one is specified.
-                          if (charCode == 45) {
-                            isSigned = true;
-                            charCode = source.charCodeAt(++Index);
-                          }
-                          // Parse an integer or floating-point value.
-                          if (charCode >= 48 && charCode <= 57) {
-                            // Leading zeroes are interpreted as octal literals.
-                            if (charCode == 48 && ((charCode = source.charCodeAt(Index + 1)), charCode >= 48 && charCode <= 57)) {
-                              // Illegal octal literal.
-                              abort();
-                            }
-                            isSigned = false;
-                            // Parse the integer component.
-                            for (; Index < length && ((charCode = source.charCodeAt(Index)), charCode >= 48 && charCode <= 57); Index++);
-                            // Floats cannot contain a leading decimal point; however, this
-                            // case is already accounted for by the parser.
-                            if (source.charCodeAt(Index) == 46) {
-                              position = ++Index;
-                              // Parse the decimal component.
-                              for (; position < length && ((charCode = source.charCodeAt(position)), charCode >= 48 && charCode <= 57); position++);
-                              if (position == Index) {
-                                // Illegal trailing decimal.
-                                abort();
-                              }
-                              Index = position;
-                            }
-                            // Parse exponents. The `e` denoting the exponent is
-                            // case-insensitive.
-                            charCode = source.charCodeAt(Index);
-                            if (charCode == 101 || charCode == 69) {
-                              charCode = source.charCodeAt(++Index);
-                              // Skip past the sign following the exponent, if one is
-                              // specified.
-                              if (charCode == 43 || charCode == 45) {
-                                Index++;
-                              }
-                              // Parse the exponential component.
-                              for (position = Index; position < length && ((charCode = source.charCodeAt(position)), charCode >= 48 && charCode <= 57); position++);
-                              if (position == Index) {
-                                // Illegal empty exponent.
-                                abort();
-                              }
-                              Index = position;
-                            }
-                            // Coerce the parsed value to a JavaScript number.
-                            return +source.slice(begin, Index);
-                          }
-                          // A negative sign may only precede numbers.
-                          if (isSigned) {
-                            abort();
-                          }
-                          // `true`, `false`, and `null` literals.
-                          if (source.slice(Index, Index + 4) == "true") {
-                            Index += 4;
-                            return true;
-                          } else if (source.slice(Index, Index + 5) == "false") {
-                            Index += 5;
-                            return false;
-                          } else if (source.slice(Index, Index + 4) == "null") {
-                            Index += 4;
-                            return null;
-                          }
-                          // Unrecognized token.
-                          abort();
-                      }
-                    }
-                    // Return the sentinel `$` character if the parser has reached the end
-                    // of the source string.
-                    return "$";
-                  };
-
-                  // Internal: Parses a JSON `value` token.
-                  var get = function(value) {
-                    var results, hasMembers;
-                    if (value == "$") {
-                      // Unexpected end of input.
-                      abort();
-                    }
-                    if (typeof value == "string") {
-                      if ((charIndexBuggy ? value.charAt(0) : value[0]) == "@") {
-                        // Remove the sentinel `@` character.
-                        return value.slice(1);
-                      }
-                      // Parse object and array literals.
-                      if (value == "[") {
-                        // Parses a JSON array, returning a new JavaScript array.
-                        results = [];
-                        for (;; hasMembers || (hasMembers = true)) {
-                          value = lex();
-                          // A closing square bracket marks the end of the array literal.
-                          if (value == "]") {
-                            break;
-                          }
-                          // If the array literal contains elements, the current token
-                          // should be a comma separating the previous element from the
-                          // next.
-                          if (hasMembers) {
-                            if (value == ",") {
-                              value = lex();
-                              if (value == "]") {
-                                // Unexpected trailing `,` in array literal.
-                                abort();
-                              }
-                            } else {
-                              // A `,` must separate each array element.
-                              abort();
-                            }
-                          }
-                          // Elisions and leading commas are not permitted.
-                          if (value == ",") {
-                            abort();
-                          }
-                          results.push(get(value));
-                        }
-                        return results;
-                      } else if (value == "{") {
-                        // Parses a JSON object, returning a new JavaScript object.
-                        results = {};
-                        for (;; hasMembers || (hasMembers = true)) {
-                          value = lex();
-                          // A closing curly brace marks the end of the object literal.
-                          if (value == "}") {
-                            break;
-                          }
-                          // If the object literal contains members, the current token
-                          // should be a comma separator.
-                          if (hasMembers) {
-                            if (value == ",") {
-                              value = lex();
-                              if (value == "}") {
-                                // Unexpected trailing `,` in object literal.
-                                abort();
-                              }
-                            } else {
-                              // A `,` must separate each object member.
-                              abort();
-                            }
-                          }
-                          // Leading commas are not permitted, object property names must be
-                          // double-quoted strings, and a `:` must separate each property
-                          // name and value.
-                          if (value == "," || typeof value != "string" || (charIndexBuggy ? value.charAt(0) : value[0]) != "@" || lex() != ":") {
-                            abort();
-                          }
-                          results[value.slice(1)] = get(lex());
-                        }
-                        return results;
-                      }
-                      // Unexpected token encountered.
-                      abort();
-                    }
-                    return value;
-                  };
-
-                  // Internal: Updates a traversed object member.
-                  var update = function(source, property, callback) {
-                    var element = walk(source, property, callback);
-                    if (element === undef) {
-                      delete source[property];
-                    } else {
-                      source[property] = element;
-                    }
-                  };
-
-                  // Internal: Recursively traverses a parsed JSON object, invoking the
-                  // `callback` function for each value. This is an implementation of the
-                  // `Walk(holder, name)` operation defined in ES 5.1 section 15.12.2.
-                  var walk = function(source, property, callback) {
-                    var value = source[property],
-                      length;
-                    if (typeof value == "object" && value) {
-                      // `forEach` can't be used to traverse an array in Opera <= 8.54
-                      // because its `Object#hasOwnProperty` implementation returns `false`
-                      // for array indices (e.g., `![1, 2, 3].hasOwnProperty("0")`).
-                      if (getClass.call(value) == arrayClass) {
-                        for (length = value.length; length--;) {
-                          update(value, length, callback);
-                        }
-                      } else {
-                        forEach(value, function(property) {
-                          update(value, property, callback);
-                        });
-                      }
-                    }
-                    return callback.call(source, property, value);
-                  };
-
-                  // Public: `JSON.parse`. See ES 5.1 section 15.12.2.
-                  exports.parse = function(source, callback) {
-                    var result, value;
-                    Index = 0;
-                    Source = "" + source;
-                    result = get(lex());
-                    // If a JSON string contains multiple tokens, it is invalid.
-                    if (lex() != "$") {
-                      abort();
-                    }
-                    // Reset the parser state.
-                    Index = Source = null;
-                    return callback && getClass.call(callback) == functionClass ? walk((value = {}, value[""] = result, value), "", callback) : result;
-                  };
-                }
-              }
-
-              exports["runInContext"] = runInContext;
-              return exports;
-            }
-
-            if (freeExports && !isLoader) {
-              // Export for CommonJS environments.
-              runInContext(root, freeExports);
-            } else {
-              // Export for web browsers and JavaScript engines.
-              var nativeJSON = root.JSON,
-                previousJSON = root["JSON3"],
-                isRestored = false;
-
-              var JSON3 = runInContext(root, (root["JSON3"] = {
-                // Public: Restores the original value of the global `JSON` object and
-                // returns a reference to the `JSON3` object.
-                "noConflict": function() {
-                  if (!isRestored) {
-                    isRestored = true;
-                    root.JSON = nativeJSON;
-                    root["JSON3"] = previousJSON;
-                    nativeJSON = previousJSON = null;
-                  }
-                  return JSON3;
-                }
-              }));
-
-              root.JSON = {
-                "parse": JSON3.parse,
-                "stringify": JSON3.stringify
-              };
-            }
-
-            // Export for asynchronous module loaders.
-            if (isLoader) {
-              define(function() {
-                return JSON3;
-              });
-            }
-          }).call(this);
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {}],
-      59: [function(require, module, exports) {
-        'use strict';
-
-        var has = Object.prototype.hasOwnProperty;
-
-        /**
-         * Simple query string parser.
-         *
-         * @param {String} query The query string that needs to be parsed.
-         * @returns {Object}
-         * @api public
-         */
-        function querystring(query) {
-          var parser = /([^=?&]+)=?([^&]*)/g,
-            result = {},
-            part;
-
-          //
-          // Little nifty parsing hack, leverage the fact that RegExp.exec increments
-          // the lastIndex property so we can continue executing this loop until we've
-          // parsed all results.
-          //
-          for (; part = parser.exec(query); result[decodeURIComponent(part[1])] = decodeURIComponent(part[2]));
-
-          return result;
-        }
-
-        /**
-         * Transform a query string to an object.
-         *
-         * @param {Object} obj Object that should be transformed.
-         * @param {String} prefix Optional prefix.
-         * @returns {String}
-         * @api public
-         */
-        function querystringify(obj, prefix) {
-          prefix = prefix || '';
-
-          var pairs = [];
-
-          //
-          // Optionally prefix with a '?' if needed
-          //
-          if ('string' !== typeof prefix) prefix = '?';
-
-          for (var key in obj) {
-            if (has.call(obj, key)) {
-              pairs.push(encodeURIComponent(key) + '=' + encodeURIComponent(obj[key]));
-            }
-          }
-
-          return pairs.length ? prefix + pairs.join('&') : '';
-        }
-
-        //
-        // Expose the module.
-        //
-        exports.stringify = querystringify;
-        exports.parse = querystring;
-
-      }, {}],
-      60: [function(require, module, exports) {
-        'use strict';
-
-        /**
-         * Check if we're required to add a port number.
-         *
-         * @see https://url.spec.whatwg.org/#default-port
-         * @param {Number|String} port Port number we need to check
-         * @param {String} protocol Protocol we need to check against.
-         * @returns {Boolean} Is it a default port for the given protocol
-         * @api private
-         */
-        module.exports = function required(port, protocol) {
-          protocol = protocol.split(':')[0];
-          port = +port;
-
-          if (!port) return false;
-
-          switch (protocol) {
-            case 'http':
-            case 'ws':
-              return port !== 80;
-
-            case 'https':
-            case 'wss':
-              return port !== 443;
-
-            case 'ftp':
-              return port !== 21;
-
-            case 'gopher':
-              return port !== 70;
-
-            case 'file':
-              return false;
-          }
-
-          return port !== 0;
-        };
-
-      }, {}],
-      61: [function(require, module, exports) {
-        'use strict';
-
-        var required = require('requires-port'),
-          lolcation = require('./lolcation'),
-          qs = require('querystringify'),
-          protocolre = /^([a-z][a-z0-9.+-]*:)?(\/\/)?([\S\s]*)/i;
-
-        /**
-         * These are the parse rules for the URL parser, it informs the parser
-         * about:
-         *
-         * 0. The char it Needs to parse, if it's a string it should be done using
-         *    indexOf, RegExp using exec and NaN means set as current value.
-         * 1. The property we should set when parsing this value.
-         * 2. Indication if it's backwards or forward parsing, when set as number it's
-         *    the value of extra chars that should be split off.
-         * 3. Inherit from location if non existing in the parser.
-         * 4. `toLowerCase` the resulting value.
-         */
-        var rules = [
-          ['#', 'hash'], // Extract from the back.
-          ['?', 'query'], // Extract from the back.
-          ['/', 'pathname'], // Extract from the back.
-          ['@', 'auth', 1], // Extract from the front.
-          [NaN, 'host', undefined, 1, 1], // Set left over value.
-          [/:(\d+)$/, 'port', undefined, 1], // RegExp the back.
-          [NaN, 'hostname', undefined, 1, 1] // Set left over.
-        ];
-
-        /**
-         * @typedef ProtocolExtract
-         * @type Object
-         * @property {String} protocol Protocol matched in the URL, in lowercase.
-         * @property {Boolean} slashes `true` if protocol is followed by "//", else `false`.
-         * @property {String} rest Rest of the URL that is not part of the protocol.
-         */
-
-        /**
-         * Extract protocol information from a URL with/without double slash ("//").
-         *
-         * @param {String} address URL we want to extract from.
-         * @return {ProtocolExtract} Extracted information.
-         * @api private
-         */
-        function extractProtocol(address) {
-          var match = protocolre.exec(address);
-
-          return {
-            protocol: match[1] ? match[1].toLowerCase() : '',
-            slashes: !!match[2],
-            rest: match[3]
-          };
-        }
-
-        /**
-         * Resolve a relative URL pathname against a base URL pathname.
-         *
-         * @param {String} relative Pathname of the relative URL.
-         * @param {String} base Pathname of the base URL.
-         * @return {String} Resolved pathname.
-         * @api private
-         */
-        function resolve(relative, base) {
-          var path = (base || '/').split('/').slice(0, -1).concat(relative.split('/')),
-            i = path.length,
-            last = path[i - 1],
-            unshift = false,
-            up = 0;
-
-          while (i--) {
-            if (path[i] === '.') {
-              path.splice(i, 1);
-            } else if (path[i] === '..') {
-              path.splice(i, 1);
-              up++;
-            } else if (up) {
-              if (i === 0) unshift = true;
-              path.splice(i, 1);
-              up--;
-            }
-          }
-
-          if (unshift) path.unshift('');
-          if (last === '.' || last === '..') path.push('');
-
-          return path.join('/');
-        }
-
-        /**
-         * The actual URL instance. Instead of returning an object we've opted-in to
-         * create an actual constructor as it's much more memory efficient and
-         * faster and it pleases my OCD.
-         *
-         * @constructor
-         * @param {String} address URL we want to parse.
-         * @param {Object|String} location Location defaults for relative paths.
-         * @param {Boolean|Function} parser Parser for the query string.
-         * @api public
-         */
-        function URL(address, location, parser) {
-          if (!(this instanceof URL)) {
-            return new URL(address, location, parser);
-          }
-
-          var relative, extracted, parse, instruction, index, key, instructions = rules.slice(),
-            type = typeof location,
-            url = this,
-            i = 0;
-
-          //
-          // The following if statements allows this module two have compatibility with
-          // 2 different API:
-          //
-          // 1. Node.js's `url.parse` api which accepts a URL, boolean as arguments
-          //    where the boolean indicates that the query string should also be parsed.
-          //
-          // 2. The `URL` interface of the browser which accepts a URL, object as
-          //    arguments. The supplied object will be used as default values / fall-back
-          //    for relative paths.
-          //
-          if ('object' !== type && 'string' !== type) {
-            parser = location;
-            location = null;
-          }
-
-          if (parser && 'function' !== typeof parser) parser = qs.parse;
-
-          location = lolcation(location);
-
-          //
-          // Extract protocol information before running the instructions.
-          //
-          extracted = extractProtocol(address || '');
-          relative = !extracted.protocol && !extracted.slashes;
-          url.slashes = extracted.slashes || relative && location.slashes;
-          url.protocol = extracted.protocol || location.protocol || '';
-          address = extracted.rest;
-
-          //
-          // When the authority component is absent the URL starts with a path
-          // component.
-          //
-          if (!extracted.slashes) instructions[2] = [/(.*)/, 'pathname'];
-
-          for (; i < instructions.length; i++) {
-            instruction = instructions[i];
-            parse = instruction[0];
-            key = instruction[1];
-
-            if (parse !== parse) {
-              url[key] = address;
-            } else if ('string' === typeof parse) {
-              if (~(index = address.indexOf(parse))) {
-                if ('number' === typeof instruction[2]) {
-                  url[key] = address.slice(0, index);
-                  address = address.slice(index + instruction[2]);
-                } else {
-                  url[key] = address.slice(index);
-                  address = address.slice(0, index);
-                }
-              }
-            } else if ((index = parse.exec(address))) {
-              url[key] = index[1];
-              address = address.slice(0, index.index);
-            }
-
-            url[key] = url[key] || (
-              relative && instruction[3] ? location[key] || '' : ''
-            );
-
-            //
-            // Hostname, host and protocol should be lowercased so they can be used to
-            // create a proper `origin`.
-            //
-            if (instruction[4]) url[key] = url[key].toLowerCase();
-          }
-
-          //
-          // Also parse the supplied query string in to an object. If we're supplied
-          // with a custom parser as function use that instead of the default build-in
-          // parser.
-          //
-          if (parser) url.query = parser(url.query);
-
-          //
-          // If the URL is relative, resolve the pathname against the base URL.
-          //
-          if (
-            relative &&
-            location.slashes &&
-            url.pathname.charAt(0) !== '/' &&
-            (url.pathname !== '' || location.pathname !== '')
-          ) {
-            url.pathname = resolve(url.pathname, location.pathname);
-          }
-
-          //
-          // We should not add port numbers if they are already the default port number
-          // for a given protocol. As the host also contains the port number we're going
-          // override it with the hostname which contains no port number.
-          //
-          if (!required(url.port, url.protocol)) {
-            url.host = url.hostname;
-            url.port = '';
-          }
-
-          //
-          // Parse down the `auth` for the username and password.
-          //
-          url.username = url.password = '';
-          if (url.auth) {
-            instruction = url.auth.split(':');
-            url.username = instruction[0] || '';
-            url.password = instruction[1] || '';
-          }
-
-          url.origin = url.protocol && url.host && url.protocol !== 'file:' ?
-            url.protocol + '//' + url.host :
-            'null';
-
-          //
-          // The href is just the compiled result.
-          //
-          url.href = url.toString();
-        }
-
-        /**
-         * This is convenience method for changing properties in the URL instance to
-         * insure that they all propagate correctly.
-         *
-         * @param {String} part          Property we need to adjust.
-         * @param {Mixed} value          The newly assigned value.
-         * @param {Boolean|Function} fn  When setting the query, it will be the function
-         *                               used to parse the query.
-         *                               When setting the protocol, double slash will be
-         *                               removed from the final url if it is true.
-         * @returns {URL}
-         * @api public
-         */
-        function set(part, value, fn) {
-          var url = this;
-
-          switch (part) {
-            case 'query':
-              if ('string' === typeof value && value.length) {
-                value = (fn || qs.parse)(value);
-              }
-
-              url[part] = value;
-              break;
-
-            case 'port':
-              url[part] = value;
-
-              if (!required(value, url.protocol)) {
-                url.host = url.hostname;
-                url[part] = '';
-              } else if (value) {
-                url.host = url.hostname + ':' + value;
-              }
-
-              break;
-
-            case 'hostname':
-              url[part] = value;
-
-              if (url.port) value += ':' + url.port;
-              url.host = value;
-              break;
-
-            case 'host':
-              url[part] = value;
-
-              if (/:\d+$/.test(value)) {
-                value = value.split(':');
-                url.port = value.pop();
-                url.hostname = value.join(':');
-              } else {
-                url.hostname = value;
-                url.port = '';
-              }
-
-              break;
-
-            case 'protocol':
-              url.protocol = value.toLowerCase();
-              url.slashes = !fn;
-              break;
-
-            case 'pathname':
-              url.pathname = value.length && value.charAt(0) !== '/' ? '/' + value : value;
-
-              break;
-
-            default:
-              url[part] = value;
-          }
-
-          for (var i = 0; i < rules.length; i++) {
-            var ins = rules[i];
-
-            if (ins[4]) url[ins[1]] = url[ins[1]].toLowerCase();
-          }
-
-          url.origin = url.protocol && url.host && url.protocol !== 'file:' ?
-            url.protocol + '//' + url.host :
-            'null';
-
-          url.href = url.toString();
-
-          return url;
-        };
-
-        /**
-         * Transform the properties back in to a valid and full URL string.
-         *
-         * @param {Function} stringify Optional query stringify function.
-         * @returns {String}
-         * @api public
-         */
-        function toString(stringify) {
-          if (!stringify || 'function' !== typeof stringify) stringify = qs.stringify;
-
-          var query, url = this,
-            protocol = url.protocol;
-
-          if (protocol && protocol.charAt(protocol.length - 1) !== ':') protocol += ':';
-
-          var result = protocol + (url.slashes ? '//' : '');
-
-          if (url.username) {
-            result += url.username;
-            if (url.password) result += ':' + url.password;
-            result += '@';
-          }
-
-          result += url.host + url.pathname;
-
-          query = 'object' === typeof url.query ? stringify(url.query) : url.query;
-          if (query) result += '?' !== query.charAt(0) ? '?' + query : query;
-
-          if (url.hash) result += url.hash;
-
-          return result;
-        }
-
-        URL.prototype = {
-          set: set,
-          toString: toString
-        };
-
-        //
-        // Expose the URL parser and some additional properties that might be useful for
-        // others or testing.
-        //
-        URL.extractProtocol = extractProtocol;
-        URL.location = lolcation;
-        URL.qs = qs;
-
-        module.exports = URL;
-
-      }, {
-        "./lolcation": 62,
-        "querystringify": 59,
-        "requires-port": 60
-      }],
-      62: [function(require, module, exports) {
-        (function(global) {
-          'use strict';
-
-          var slashes = /^[A-Za-z][A-Za-z0-9+-.]*:\/\//;
-
-          /**
-           * These properties should not be copied or inherited from. This is only needed
-           * for all non blob URL's as a blob URL does not include a hash, only the
-           * origin.
-           *
-           * @type {Object}
-           * @private
-           */
-          var ignore = {
-              hash: 1,
-              query: 1
-            },
-            URL;
-
-          /**
-           * The location object differs when your code is loaded through a normal page,
-           * Worker or through a worker using a blob. And with the blobble begins the
-           * trouble as the location object will contain the URL of the blob, not the
-           * location of the page where our code is loaded in. The actual origin is
-           * encoded in the `pathname` so we can thankfully generate a good "default"
-           * location from it so we can generate proper relative URL's again.
-           *
-           * @param {Object|String} loc Optional default location object.
-           * @returns {Object} lolcation object.
-           * @api public
-           */
-          module.exports = function lolcation(loc) {
-            loc = loc || global.location || {};
-            URL = URL || require('./');
-
-            var finaldestination = {},
-              type = typeof loc,
-              key;
-
-            if ('blob:' === loc.protocol) {
-              finaldestination = new URL(unescape(loc.pathname), {});
-            } else if ('string' === type) {
-              finaldestination = new URL(loc, {});
-              for (key in ignore) delete finaldestination[key];
-            } else if ('object' === type) {
-              for (key in loc) {
-                if (key in ignore) continue;
-                finaldestination[key] = loc[key];
-              }
-
-              if (finaldestination.slashes === undefined) {
-                finaldestination.slashes = slashes.test(loc.href);
-              }
-            }
-
-            return finaldestination;
-          };
-
-        }).call(this, typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-
-      }, {
-        "./": 61
-      }]
-    }, {}, [1])(1)
-  });
-
-
-  //# sourceMappingURL=sockjs.js.map
-
-  var SockJSImpl = window.SockJS;
-
-  var SockJSWrapper = function(url, options) {
-    var self = this;
-
-    options = options || {};
-    self.reconnectTimerID = null;
-    self.reconnectEnabled = false;
-    self.reconnectAttempts = 0;
-    self.firstMessage = true;
-    self.maxReconnectAttempts = options.maxReconnectAttempts || Infinity;
-    self.reconnectInterval = options.reconnectInterval || 5000;
-
-
-    var setupSockJSConnection = function() {
-        self.sock = SockJSImpl(url, null, options);
-        self.sock.onopen = function() {
-            self.firstMessage = true;
-            if (self.reconnectTimerID) {
-              self.reconnectAttempts = 0;
-              // fire separate event for reconnects
-              // consistent behavior with adding handlers onopen
-              self.onreopen && self.onreopen();
-            }
-        };
-
-        self.sock.onclose = function (e) {
-            if (self.reconnectEnabled) {
-              if (self.reconnectAttempts < self.maxReconnectAttempts) {
-                self.onreconnect && self.onreconnect();
-                self.sock = null;
-                // set id so users can cancel
-                self.reconnectTimerID = setTimeout(setupSockJSConnection, self.reconnectInterval);
-                ++self.reconnectAttempts;
-              } else {
-                // notify error
-                var e = new Event('reconnectionError');
-                e.transport = self.sock.transport;
-                self.onerror && self.onerror(e);
-              }
-            } else {
-                self.onclose && self.onclose(e);
-            }
-        };
-        self.sock.onmessage = function(e) {
-            if (self.firstMessage) {
-                self.firstMessage = false;
-                // TODO: handle first message?
-                self.onopen && self.onopen();
-            } else {
-                self.onmessage && self.onmessage(e);
-            }
-        };
-        self.sock.onerror = function(e) {
-            self.onerror && self.onerror(e);
-        }
-    };
-
-    setupSockJSConnection();
-
-  };
-
-    SockJSWrapper.prototype.close = function () {
-        this.enableReconnect(false);
-        this.sock.close();
-    };
-    SockJSWrapper.prototype.send = function(message) {
-        if (this.sock) {
-            this.sock.send(message);
-        } else {
-            throw new Error('SockJS not initialized');
-        }
-    };
-    SockJSWrapper.prototype.enableReconnect = function (enable) {
-        var self = this;
-
-        self.reconnectEnabled = enable;
-        if (!enable && self.reconnectTimerID) {
-          clearTimeout(self.reconnectTimerID);
-          self.reconnectTimerID = null;
-          self.reconnectAttempts = 0;
-        }
-    };
-    SockJSWrapper.prototype.getTransport = function () {
-        return this.sock ? this.sock.transport : null;
-    }
-    SockJSWrapper.prototype.getReadyState = function () {
-        return this.sock ? this.sock.readyState : SockJSImpl.CONNECTING;
-    };
-
-  this.SockJS = {
-    connect: function(config) {
-        var sock = new SockJSWrapper(config.url, config);
-        sock.enableReconnect(parseInt(config.reconnectInterval) > 0);
-        sock.onopen = config.onOpen;
-        sock.onmessage = config.onMessage;
-        sock.onclose = config.onClose;
-        sock.onerror = config.onError
-        sock.onreconnect = config.onReconnect;
-        sock.onreopen = config.onReopen;
-        return sock;
-    },
-  }
-
-
-}).call(window.vaadinPush);
